% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_assess.R
\name{all_cohort_assessments}
\alias{all_cohort_assessments}
\title{A default list of assessments to perform for each package}
\usage{
all_cohort_assessments()
}
\value{
a list of assess_* functions exported from riskmetric
}
\description{
A default list of assessments to perform for each package
}
