% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_metric.R
\name{as_cohort_metric}
\alias{as_cohort_metric}
\title{Convert an object to a \code{cohort_metric}}
\usage{
as_cohort_metric(x, class = c())
}
\arguments{
\item{x}{data to store as a \code{cohort_metric}}

\item{class}{a subclass to differentiate the \code{cohort_metric} object}
}
\value{
a \code{cohort_metric} object
}
\description{
Convert an object to a \code{cohort_metric}
}
