% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_metric.R
\name{as_cohort_metric_condition}
\alias{as_cohort_metric_condition}
\title{A cohort_metric subclass for general metric evaluation conditions}
\usage{
as_cohort_metric_condition(x, ..., subclass = c())
}
\arguments{
\item{x}{an object to wrap in a \code{cohort_metric_condition} class}

\item{...}{additional arguments added as attributes to object \code{x}}

\item{subclass}{an optional subclass of \code{cohort_metric_condition} to
include}
}
\value{
an object after wrap \code{cohort_metric_condition} class.
}
\description{
A cohort_metric subclass for general metric evaluation conditions
}
\keyword{internal}
