% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_metric.R
\name{as_cohort_metric_error}
\alias{as_cohort_metric_error}
\title{A subclass wrapping an error with an additional parent class}
\usage{
as_cohort_metric_error(error)
}
\arguments{
\item{error}{an error condition object to capture}
}
\value{
an error condition object after wrap \code{cohort_metric_error} class.
}
\description{
A subclass wrapping an error with an additional parent class
}
\keyword{internal}
