% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_metric.R
\name{as_pkg_metric}
\alias{as_pkg_metric}
\title{Convert an object to a \code{pkg_metric}}
\usage{
as_pkg_metric(x, class = c())
}
\arguments{
\item{x}{data to store as a \code{pkg_metric}}

\item{class}{a subclass to differentiate the \code{pkg_metric} object}
}
\description{
Convert an object to a \code{pkg_metric}
}
