% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_metric_condition.R
\name{as_pkg_metric_condition}
\alias{as_pkg_metric_condition}
\title{A pkg_metric subclass for general metric evaluation conditions}
\usage{
as_pkg_metric_condition(x, ..., subclass = c())
}
\arguments{
\item{message}{an optional message explaining why a metric is not applicable.}
}
\description{
A pkg_metric subclass for general metric evaluation conditions
}
