% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_metric_error.R
\name{as_pkg_metric_error}
\alias{as_pkg_metric_error}
\title{A subclass wrapping an error with an additional parent class}
\usage{
as_pkg_metric_error(error)
}
\arguments{
\item{error}{an error condition object to capture}

\item{superclass}{an optional superclass for the pkg_metric subclass (more
specific than a pkg_metric, but less specific than the error subclass)}
}
\description{
A subclass wrapping an error with an additional parent class
}
