% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_metric_na.R
\name{as_pkg_metric_na}
\alias{as_pkg_metric_na}
\title{A pkg_metric subclass for when metrics are explicitly not applicable}
\usage{
as_pkg_metric_na(x, message = NULL)
}
\arguments{
\item{x}{a \code{pkg_metric} object to wrap in a \code{pkg_metric_na}
subclass}

\item{message}{an optional message explaining why a metric is not applicable.}
}
\description{
A pkg_metric subclass for when metrics are explicitly not applicable
}
\keyword{internal}
