% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_metric_todo.R
\name{as_pkg_metric_todo}
\alias{as_pkg_metric_todo}
\title{A pkg_metric subclass for when pkg_metrics have not yet been implemented}
\usage{
as_pkg_metric_todo(x, message = NULL)
}
\arguments{
\item{message}{an optional message directing users and potential contributors
toward any ongoing work or first steps toward development.}
}
\description{
A pkg_metric subclass for when pkg_metrics have not yet been implemented
}
