% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_class.R
\name{as_pkg_ref}
\alias{as_pkg_ref}
\title{Convert into a package object}
\usage{
as_pkg_ref(x, ...)
}
\arguments{
\item{x}{A singular \code{character} value, \code{character vector} or
\code{list} of \code{character} values of package names or source code
directory paths.}

\item{...}{Additional arguments passed to methods.}
}
\value{
When a single value is provided, a single \code{pkg_ref} object is
  returned, possibly with a subclass based on where the package was found. If
  a \code{vector} or \code{list} is provided, a \code{list_of_pkg_ref} object
  constructed with \code{\link[vctrs]{list_of}} is returned, which can be
  considered analogous to a \code{list}. See 'Details' for further
  information about \code{pkg_ref} subclasses.
}
\description{
Convert into a package object
}
\seealso{
Other pkg_ref: 
\code{\link{pkg_ref}()}
}
\concept{pkg_ref}
