% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{assess}
\alias{assess}
\title{Function for applying assess_* family of functions to packages}
\usage{
assess(x, assessments = all_assessments(), ...,
  error = assessment_error_as_warning)
}
\description{
Function for applying assess_* family of functions to packages
}
