% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_remote_checks.R
\name{assess_cran_checks}
\alias{assess_cran_checks}
\title{Assess package checks from CRAN/Bioc or R CMD check}
\usage{
assess_cran_checks(x, ...)
}
\description{
Assess package checks from CRAN/Bioc or R CMD check
}
