% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_dependencies.R
\name{assess_dependencies}
\alias{assess_dependencies}
\title{Assessment of dependency footprint for a specific package}
\usage{
assess_dependencies(x, ...)
}
\description{
Only Depends, Imports and LinkingTo dependencies are assessed because
they are required
}
\details{
The more packages a package relies on the more chances for errors exist.
}
