% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_downloads.R
\name{assess_downloads_1yr}
\alias{assess_downloads_1yr}
\title{Assess a package for the number of downloads in the past year}
\usage{
assess_downloads_1yr(x, ...)
}
\description{
Assess a package for the number of downloads in the past year
}
\details{
The more times a package has been downloaded the more extensive the user testing and the greater chance there is of someone finding a bug and logging it.
}
