% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_export_help.R
\name{assess_export_help}
\alias{assess_export_help}
\title{Assess a package for availability of documentation for exported values}
\usage{
assess_export_help(x, ...)
}
\arguments{
\item{x}{a \code{pkg_ref} package reference object}

\item{...}{additional arguments passed on to S3 methods, rarely used}
}
\value{
a \code{pkg_metric} containing a logical vector indicating existence of documentation for each namespace export
}
\description{
Assess a package for availability of documentation for exported values
}
\examples{
\dontrun{
assess_export_help(pkg_ref("riskmetric"))
}
}
\seealso{
\code{\link{metric_score.pkg_metric_export_help}}
}
