% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{assess_family_roxygen}
\alias{assess_family_roxygen}
\title{Helper for creating a roxygen header from template for assess_* functions}
\usage{
assess_family_roxygen(name, return_type = "an atomic assessment result")
}
\description{
Helper for creating a roxygen header from template for assess_* functions
}
\examples{
\dontrun{
  #' @eval assess_family_roxygen(
      "has_news",
      "an integer value indicating the number of discovered NEWS files")
}

}
