% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_bug_reports_url.R
\name{assess_has_bug_reports_url}
\alias{assess_has_bug_reports_url}
\title{Assess a package for the presence of a url field where bugs can be reported.}
\usage{
assess_has_bug_reports_url(x, ...)
}
\description{
Assess a package for the presence of a url field where bugs can be reported.
}
