% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_dependency_conflict.R
\name{assess_has_dependency_conflict}
\alias{assess_has_dependency_conflict}
\title{Assess a cohort for the presence of dependency conflicts}
\usage{
assess_has_dependency_conflict(x, ...)
}
\description{
Assess a cohort for the presence of dependency conflicts
}
