% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_maintainer.R
\name{assess_has_maintainer}
\alias{assess_has_maintainer}
\title{Assess a package for an associated maintainer}
\usage{
assess_has_maintainer(x, ...)
}
\description{
Assess a package for an associated maintainer
}
