% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_news.R
\name{assess_has_news}
\alias{assess_has_news}
\title{Assess a package for the presence of a NEWS file}
\usage{
assess_has_news(x, ...)
}
\arguments{
\item{x}{a packge reference object}

\item{...}{additional arguments unused}
}
\value{
a \code{pkg_metric} object
}
\description{
Assess a package for the presence of a NEWS file
}
