% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_source_control.R
\name{assess_has_source_control}
\alias{assess_has_source_control}
\title{Assess a package for an associated source control url}
\usage{
assess_has_source_control(x, ...)
}
\description{
Assess a package for an associated source control url
}
