% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_vignettes.R
\name{assess_has_vignettes}
\alias{assess_has_vignettes}
\title{Assess a package for the presence of Vignettes files}
\usage{
assess_has_vignettes(x, ...)
}
\description{
Assess a package for the presence of Vignettes files
}
