% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_website.R
\name{assess_has_website}
\alias{assess_has_website}
\title{Assess a package for an associated website url}
\usage{
assess_has_website(x, ...)
}
\description{
Assess a package for an associated website url
}
