% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_bug_reports_closed.R
\name{assess_last_30_bugs_status}
\alias{assess_last_30_bugs_status}
\title{Assess how many recent BugReports have been closed}
\usage{
assess_last_30_bugs_status(x, ...)
}
\description{
Assess how many recent BugReports have been closed
}
