% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_license.R
\name{assess_license}
\alias{assess_license}
\title{Assess a package for an acceptable license}
\usage{
assess_license(x, ...)
}
\arguments{
\item{x}{a \code{pkg_ref} package reference object}

\item{...}{additional arguments passed on to S3 methods, rarely used}
}
\value{
a \code{pkg_metric} containing a string indicating the license under which the package is released
}
\description{
Assess a package for an acceptable license
}
\examples{
assess_license(pkg_ref("riskmetric"))
}
\seealso{
\code{\link{metric_score.pkg_metric_license}}
}
