% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_news_current.R
\name{assess_news_current}
\alias{assess_news_current}
\title{Assess a package for an up-to-date NEWS file}
\usage{
assess_news_current(x, ...)
}
\arguments{
\item{x}{a packge reference object}

\item{...}{additional arguments unused}
}
\value{
a \code{pkg_metric} object
}
\description{
Assess a package for an up-to-date NEWS file
}
