% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_r_cmd_check.R
\name{assess_r_cmd_check}
\alias{assess_r_cmd_check}
\title{Assess a package's results from running R CMD check}
\usage{
assess_r_cmd_check(x, ...)
}
\description{
Assess a package's results from running R CMD check
}
