% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_security.R
\name{assess_security}
\alias{assess_security}
\title{Assess a package for known security vulnerabilities in the OSS Index}
\usage{
assess_security(x, ...)
}
\arguments{
\item{x}{a \code{pkg_ref} package reference object}

\item{...}{additional arguments passed on to S3 methods, rarely used}
}
\value{
a \code{pkg_metric} containing Assess for any known security vulnerabilities in the OSS Index via oysteR
}
\description{
Assess a package for known security vulnerabilities in the OSS Index
}
\seealso{
\code{\link{metric_score.pkg_metric_security}}
}
