% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_metric_error.R
\name{assessment_error_as_warning}
\alias{assessment_error_as_warning}
\title{Error handler for assessments to deescalate to warning}
\usage{
assessment_error_as_warning(e, name, assessment)
}
\description{
Error handler for assessments to deescalate to warning
}
