% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache.R
\name{available_pkg_ref_fields}
\alias{available_pkg_ref_fields}
\title{A helper function for retrieving a list of available fields, identified based
on implementation of a pkg_ref_cache method for a given class.}
\usage{
available_pkg_ref_fields(x)
}
\arguments{
\item{x}{a package reference object}
}
\value{
a list of available fields implemented with a pkg_ref_cache method
}
\description{
A helper function for retrieving a list of available fields, identified based
on implementation of a pkg_ref_cache method for a given class.
}
\keyword{internal}
