% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_bug_reports.R
\name{bug_report_metadata}
\alias{bug_report_metadata}
\title{Helper for structuring bug reports}
\usage{
bug_report_metadata(bug_reports_data, x)
}
\description{
Helper for structuring bug reports
}
