% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_last_30_bugs_status.R
\name{bug_reports_status}
\alias{bug_reports_status}
\title{Retrieve statuses of bug reports}
\usage{
bug_reports_status(x, ...)
}
\arguments{
\item{x}{a \code{pkg_ref} package reference object}

\item{...}{additional arguments passed on to S3 methods, rarely used}
}
\value{
a logical vector indicating whether a bug report is closed
}
\description{
Retrieve statuses of bug reports
}
