% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache__behaviors.R
\docType{data}
\name{cache_behaviors}
\alias{cache_behaviors}
\title{List of available caching behaviors with metadata, including default and
annotations for building documentation}
\format{An object of class \code{list} of length 1.}
\usage{
cache_behaviors
}
\description{
List of available caching behaviors with metadata, including default and
annotations for building documentation
}
\keyword{datasets}
