% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capture_expr_output}
\alias{capture_expr_output}
\title{Capture side effects issued by an evaluated expression}
\usage{
capture_expr_output(expr)
}
\arguments{
\item{expr}{an expression to evaluate, capturing output events as they
are issued}
}
\description{
All messaging condition side effects are captured in the order that they
are issued.
}
