% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_assess.R
\name{cohort_assess}
\alias{cohort_assess}
\title{Apply assess_* family of functions to a cohort reference}
\usage{
cohort_assess(
  x,
  assessments = cohort_assessments(),
  ...,
  error_handler = assessment_error_empty
)
}
\arguments{
\item{x}{A single \code{\link{cohort_ref}} object or
\code{\link[tibble]{tibble}} of cohort references to assess}

\item{assessments}{A list of assessment functions to apply to each cohort
reference. By default, a list of all exported assess_* functions from the
riskmetric package.}

\item{...}{additional arguments unused}

\item{error_handler}{A function, which accepts a single parameter expecting
the raised error, which will be called if any errors occur when attempting
to apply an assessment function.}
}
\value{
A \code{list_of_cohort_metric} object
}
\description{
By default, use all \code{assess_*} functions in the \code{riskmetric}
namespace and produce a \code{\link[tibble]{tibble}} with one column per
assessment applied.
}
