% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_dependency_conflict.R
\name{cohort_metric.cohort_metric_dependency_conflict}
\alias{cohort_metric.cohort_metric_dependency_conflict}
\title{Score a package for presence of dependency conflicts}
\usage{
cohort_metric.cohort_metric_dependency_conflict(x, ...)
}
\value{
\code{1} if any dependencies are missing or conflicting, otherwise \code{0}
}
\description{
Score a package for presence of dependency conflicts
}
