% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_metric.R
\name{cohort_metric}
\alias{cohort_metric}
\title{A helper for structuring assessment return objects for dispatch with the
score function}
\usage{
cohort_metric(x = NA, ..., class = c())
}
\arguments{
\item{x}{data to store as a \code{cohort_metric}}

\item{...}{additional attributes to bind to the \code{cohort_metric} object}

\item{class}{a subclass to differentiate the \code{cohort_metric} object}
}
\value{
a \code{cohort_metric} object
}
\description{
A helper for structuring assessment return objects for dispatch with the
score function
}
