% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_ref_class.R
\name{cohort_ref}
\alias{cohort_ref}
\title{Cohort ref class}
\usage{
cohort_ref(
  x,
  library = c("base", "recommended", "installed"),
  lib.loc = "",
  includeDependencies = TRUE,
  ...
)
}
\arguments{
\item{x}{either a list of `pkg_ref`s or a `list_of_pkg_ref`}

\item{library}{one of "base", "recommended" or "installed" to use as an environment for testing cohorts}

\item{lib.loc}{path to R library}

\item{includeDependencies}{logical. whether or not to added add all pkg dependencies when creating a cohort [default: TRUE]}

\item{...}{named vector or list of paramters to be included in cohort ref creation}
}
\description{
This is a class to hold a list of pkg_refs along with extra
information about the environment the packages are to be used/installed in.
}
\concept{cohort_ref}
