% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_metric_coverage.R
\name{coverage}
\alias{coverage}
\title{Compute test coverage using covr for a RemoteReference}
\usage{
coverage(pkg_ref, ...)
}
\arguments{
\item{pkg_ref}{remote package reference to work with - must have been
downloaded already}

\item{...}{further arguments passed to methods}
}
\description{
Compute test coverage using covr for a RemoteReference
}
