% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_class_extract.R
\name{dec_mutations_count}
\alias{dec_mutations_count}
\title{decrement the number of allowed mutations}
\usage{
dec_mutations_count(x)
}
\arguments{
\item{x}{pkg_ref object to decrement mutation counter for}
}
\value{
pkg_ref object
}
\description{
decrement the number of allowed mutations
}
\keyword{internal}
