% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_class.R
\name{determine_pkg_source}
\alias{determine_pkg_source}
\title{Determine the intended source of a new package}
\usage{
determine_pkg_source(x, source, repos)
}
\arguments{
\item{x}{Package name or path to package}

\item{source}{type of source passed in `pkg_ref`}
}
\value{
one of c('pkg_source', 'pkg_install', 'pkg_cran_remote',
  'pkg_bioc_remote', 'pkg_missing')
}
\description{
Determine the intended source of a new package
}
\keyword{internal}
