% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_export_help.R
\name{dev_tips.pkg_metric_export_help}
\alias{dev_tips.pkg_metric_export_help}
\title{Provide development hints for improving exported value documentation}
\usage{
\method{dev_tips}{pkg_metric_export_help}(x, ...)
}
\arguments{
\item{x}{a pkg_metric object}

\item{...}{additional arguments unused}
}
\description{
Provide development hints for improving exported value documentation
}
