% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_tips.R
\name{dev_tips}
\alias{dev_tips}
\title{Provide tips for improving a metric}
\usage{
dev_tips(x, ...)
}
\arguments{
\item{x}{a pkg_metric object}

\item{...}{additional arguments unused}
}
\description{
Provide tips for improving a metric
}
