% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_risk.R
\docType{data}
\name{.risk_weights}
\alias{.risk_weights}
\title{Default weights to use for summarizing risk}
\format{
An object of class \code{numeric} of length 8.
}
\usage{
.risk_weights
}
\description{
Default weights to use for summarizing risk
}
\keyword{datasets}
