% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoteReference.R
\name{download_pkg_sources}
\alias{download_pkg_sources}
\title{Download source code from RemoteReference}
\usage{
download_pkg_sources(remote_ref, dest, ...)
}
\arguments{
\item{remote_ref}{remote reference}

\item{dest}{desitnation path for downloaded sources}

\item{...}{further parameters passed to methods}
}
\description{
Download source code from RemoteReference
}
