% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_examples.R
\name{examples_from_dir}
\alias{examples_from_dir}
\title{Build logical vector for Rd objects with example or usage fields discovered in a given directory}
\usage{
examples_from_dir(path, pkg)
}
\arguments{
\item{path}{a package directory path expected to contain exported objects}
}
\value{
a numeric proportion of documentation files with examples
}
\description{
Build logical vector for Rd objects with example or usage fields discovered in a given directory
}
\keyword{internal}
