% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_examples.R
\name{examples_from_pkg}
\alias{examples_from_pkg}
\title{Build logical vector for Rd objects with example or usage fields discovered in a given package}
\usage{
examples_from_pkg(pkg)
}
\arguments{
\item{pkg}{a package name expected to contain exported objects}
}
\value{
a numeric proportion of documentation files with examples
}
\description{
Build logical vector for Rd objects with example or usage fields discovered in a given package
}
\keyword{internal}
