% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_examples.R
\name{filter_rd_db}
\alias{filter_rd_db}
\title{Filter a simple database of Rd objects in a package for files with example
fields}
\usage{
filter_rd_db(rddb)
}
\arguments{
\item{rddb}{a simple database of Rd object obtained via tools::Rd_db}
}
\value{
a vector of Rd file names that have example fields
}
\description{
Filter a simple database of Rd objects in a package for files with example
fields
}
\keyword{internal}
