% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_scores.R
\name{find_all_weights}
\alias{find_all_weights}
\title{create vector of weights based on the data columms and user-given weights}
\usage{
find_all_weights(data, weights)
}
\arguments{
\item{data}{a \code{\link[tibble]{tibble}} or list of scored assessments
whose column names match those provided by riskmetric's \code{\link{assess}}
function.}

\item{weights}{a set of numeric weights to give to each score column when
calculating risk.}
}
\value{
a named vector with weights for each score column. If a
weight is not provided by the user, it will default to 1.
}
\description{
create vector of weights based on the data columms and user-given weights
}
