% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{firstS3method}
\alias{firstS3method}
\title{Find the S3 method that will be evaluated when an S3 generic is called by
an object of class \code{classes}}
\usage{
firstS3method(f, classes, envir = parent.frame())
}
\arguments{
\item{f}{character: name of the generic.}

\item{classes}{a character vector of classes used to search for the
appropriate S3 method}

\item{envir}{the \code{\link{environment}} in which the method and its
    generic are searched first.}
}
\description{
Find the S3 method that will be evaluated when an S3 generic is called by
an object of class \code{classes}
}
\keyword{internal}
