% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_metric_error.R
\name{format_assessment_message}
\alias{format_assessment_message}
\title{Assessment console printing formatter}
\usage{
format_assessment_message(e, name, assessment)
}
\arguments{
\item{e}{an error raised during a package reference assessment}

\item{name}{the name of the package whose package reference assessment raised
the error}

\item{assessment}{the name of the assessment function which raised the error}
}
\value{
a character string of formatted text to communicate the error
}
\description{
make the errors and warnings consistent with meaningful indication of what
triggered the error, including the name of the package whose reference
triggered the error while running which asesessment.
}
\keyword{internal}
