% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{get_assessment_columns}
\alias{get_assessment_columns}
\title{Helper for retrieving a list of columns which contain pkg_metric objects}
\usage{
get_assessment_columns(tbl)
}
\description{
Helper for retrieving a list of columns which contain pkg_metric objects
}
