% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_assess.R
\name{get_assessments}
\alias{get_assessments}
\title{Get a specific set of assess_* functions for pkg_assess}
\usage{
get_assessments(fxn_string = "")
}
\value{
a list of specific assess_* functions exported from riskmetric
}
\description{
Get a specific set of assess_* functions for pkg_assess
}
