% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.R
\name{get_cran_version}
\alias{get_cran_version}
\title{Get current CRAN version}
\usage{
get_cran_version(pkgname)
}
\arguments{
\item{pkgname}{a character vector of names of packages}
}
\value{
a character vector of version numbers of respective packages
}
\description{
Get current CRAN version
}
\examples{
get_cran_version(c("dplyr", "tidyverse"))

}
