% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_dependencies.R
\name{get_package_dependencies}
\alias{get_package_dependencies}
\title{Gets available packages from necessary repository and filters for
package of interest}
\usage{
get_package_dependencies(name, repo)
}
\arguments{
\item{name}{package name}

\item{repo}{package repository (e.g. CRAN or Bioconductor)}
}
\value{
Returns a data frame with two columns 1) Package names, 2) type of dependency (LinkingTo, Imports, Depends)
}
\description{
Gets available packages from necessary repository and filters for
package of interest
}
\keyword{internal}
