% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_class.R
\name{get_pkg_ref_classes}
\alias{get_pkg_ref_classes}
\title{Walk the pkg_ref class hierarchy to match a single subclass to a class path}
\usage{
get_pkg_ref_classes(x, classes = pkg_ref_class_hierarchy)
}
\arguments{
\item{x}{(`character(1L)`) A subclass, among those known in pkg_ref subclasses}

\item{classes}{(`list`) A class hierarchy, described using a named list.
Defaults to `pkg_ref_class_hierarchy`.}
}
\value{
A `character(n)` class path from `pkg_ref` down to the specified
  subclass, or `FALSE` if no path is found.
}
\description{
Walk the pkg_ref class hierarchy to match a single subclass to a class path
}
