% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{if_not_null_else}
\alias{if_not_null_else}
\alias{\%||\%}
\title{If not NULL else}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{Left-hand side}

\item{rhs}{Right-hand side

A shorthand for a common comparison}
}
\value{
an object same as \code{lhs} or \code{rhs}
}
\description{
If not NULL else
}
\keyword{internal}
