% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_class_extract.R
\name{inc_mutations_count}
\alias{inc_mutations_count}
\title{increment the number of allowed mutations}
\usage{
inc_mutations_count(x)
}
\arguments{
\item{x}{pkg_ref object to increment mutation counter for}
}
\value{
a pkg_ref object
}
\description{
increment the number of allowed mutations
}
\keyword{internal}
