% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{installed_packages}
\alias{installed_packages}
\title{Return a vector of random installed packages}
\usage{
installed_packages(...)
}
\value{
a character vector of random package names
}
\description{
Return a vector of random installed packages
}
