% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_url_subpath_of}
\alias{is_url_subpath_of}
\title{check if a url originates from a list of repo urls}
\usage{
is_url_subpath_of(url, urls)
}
\description{
check if a url originates from a list of repo urls
}
