% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_memoised.R
\name{memoise_bioc_mirrors}
\alias{memoise_bioc_mirrors}
\title{Fetch BioC Mirrors Info}
\usage{
memoise_bioc_mirrors(
  host_exists = !is.null(curl::nslookup("bioconductor.org", error = FALSE))
)
}
\arguments{
\item{host_exists}{whether bioconductor can be contacted}
}
\description{
taken from utils::chooseBioCmirror
}
