% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_memoised.R
\name{memoise_cran_mirrors}
\alias{memoise_cran_mirrors}
\title{Fetch CRAN Mirrors Info}
\usage{
memoise_cran_mirrors(all = TRUE, ..., .local = getOption("riskmetric.tests"))
}
\arguments{
\item{all}{default \code{TRUE}, passed to \code{\link{utils}[getCRANmirrors]}}

\item{...}{additional arguments passed to \code{\link{utils}[getCRANmirrors]}}

\item{.local}{an optional local directory to source the CRAN package index
from, defaulting to \code{getOption("riskmetric.tests")}, used
for isolating repository requests during testing.}
}
\description{
Fetch CRAN Mirrors Info
}
\keyword{internal}
