% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_memoised.R
\name{memoise_cran_mirros}
\alias{memoise_cran_mirros}
\title{Fetch CRAN Mirrors Info}
\usage{
memoise_cran_mirros(all = TRUE, ...,
  host_exists = !is.null(curl::nslookup("cran.r-project.org", error =
  FALSE)))
}
\arguments{
\item{all}{default \code{TRUE}, passed to \code{\link{utils}[getCRANmirrors]}}

\item{...}{additional arguments passed to \code{\link{utils}[getCRANmirrors]}}

\item{host_exists}{whether cran can be contacted}
}
\description{
Fetch CRAN Mirrors Info
}
