% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_dependency_conflict.R
\name{metric_score.cohort_metric_dependency_conflict}
\alias{metric_score.cohort_metric_dependency_conflict}
\title{Score a package for presence of dependency conflicts}
\usage{
\method{metric_score}{cohort_metric_dependency_conflict}(x, ...)
}
\arguments{
\item{x}{cohort assessment}

\item{...}{named list of arguments to adjust metric scoring. currently ignored.}
}
\value{
\code{1} if any dependencies are missing or conflicting, otherwise \code{0}
}
\description{
Score a package for presence of dependency conflicts
}
