% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_exported_namespace.R
\name{metric_score.cohort_metric_exported_namespace}
\alias{metric_score.cohort_metric_exported_namespace}
\title{Score a cohort for the number of exported objects}
\usage{
\method{metric_score}{cohort_metric_exported_namespace}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_exported_namespace} packge metric object}

\item{...}{additional arguments unused}
}
\value{
numeric value between \code{0} (high number of exported objects) and
  \code{1} (low number of exported objects)
}
\description{
Score a cohort for the number of exported objects it has; regularized
Convert the number of exported objects \code{length(x)} into a validation
score [0,1] \deqn{ 1 / (1 + exp(-0.5 * (sqrt(length(x)) + sqrt(5)))) }
}
\details{
The scoring function is the classic logistic curve \deqn{
1 / (1 + exp(-k(x-x[0])) } with a square root scale for the number of exported objects
\eqn{x = sqrt(length(x))}, sigmoid midpoint is 25 exported objects, ie. \eqn{x[0] =
sqrt(5)}, and logistic growth rate of \eqn{k = 0.25}.

\deqn{ 1 / (1 + exp(-0.25 * sqrt(length(x))-sqrt(25))) }
}
\examples{
\dontrun{metric_score(assess_exported_namespace(pkg_ref("riskmetric")))
}
}
