% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_reverse_dependencies.R
\name{metric_score.cohort_metric_reverse_dependencies}
\alias{metric_score.cohort_metric_reverse_dependencies}
\title{Metric for the assessment of reverse dependencies of a cohort.}
\usage{
\method{metric_score}{cohort_metric_reverse_dependencies}(x, ...)
}
\arguments{
\item{x}{cohort assessment}

\item{...}{named list of arguments that are currently ignored}
}
\description{
Metric for the assessment of reverse dependencies of a cohort.
}
