% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_covr_coverage.R
\name{metric_score.pkg_metric_covr_coverage}
\alias{metric_score.pkg_metric_covr_coverage}
\title{Score a package for unit test coverage}
\usage{
\method{metric_score}{pkg_metric_covr_coverage}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_covr_coverage} packge metric object}

\item{...}{additional arguments unused}
}
\value{
A \code{numeric}
}
\description{
Returns the overall test coverage from a covr coverage report
}
\examples{
\dontrun{metric_score(assess_covr_coverage(pkg_ref("riskmetric")))
}
}
