% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_export_help.R
\name{metric_score.pkg_metric_export_help}
\alias{metric_score.pkg_metric_export_help}
\title{Score a package for availability of documentation for exported values}
\usage{
\method{metric_score}{pkg_metric_export_help}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_export_help} packge metric object}

\item{...}{additional arguments unused}
}
\value{
\code{1} if any NEWS files are found, otherwise \code{0}
}
\description{
Coerce a logical vector indicating availability of export documentation
}
\examples{
metric_score(assess_export_help(pkg_ref("riskmetric")))
}
\seealso{
Other \code{metric_score.*} functions: 
\code{\link{metric_score.pkg_metric_covr_coverage}()},
\code{\link{metric_score.pkg_metric_downloads_1yr}()},
\code{\link{metric_score.pkg_metric_has_bug_reports_url}()},
\code{\link{metric_score.pkg_metric_has_maintainer}()},
\code{\link{metric_score.pkg_metric_has_news}()},
\code{\link{metric_score.pkg_metric_has_source_control}()},
\code{\link{metric_score.pkg_metric_has_vignettes}()},
\code{\link{metric_score.pkg_metric_has_website}()},
\code{\link{metric_score.pkg_metric_last_30_bugs_status}()},
\code{\link{metric_score.pkg_metric_license}()},
\code{\link{metric_score.pkg_metric_news_current}()},
\code{\link{metric_score.pkg_metric_r_cmd_check}()}
}
\concept{\code{metric_score.*} functions}
