% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_exported_namespace.R
\name{metric_score.pkg_metric_exported_namespace}
\alias{metric_score.pkg_metric_exported_namespace}
\title{Score a package for the number of exported objects}
\usage{
\method{metric_score}{pkg_metric_exported_namespace}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_exported_namespace} packge metric object}

\item{...}{additional arguments unused}
}
\value{
numeric value
}
\description{
Count the number of exported objects (excluding S3Methods) and divide by 100
}
\examples{
\dontrun{metric_score(assess_exported_namespace(pkg_ref("riskmetric")))
}
}
