% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_bug_reports_url.R
\name{metric_score.pkg_metric_has_bug_reports_url}
\alias{metric_score.pkg_metric_has_bug_reports_url}
\title{Score a package for the presence of a bug report url}
\usage{
\method{metric_score}{pkg_metric_has_bug_reports_url}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_has_bug_reports_url} packge metric object}

\item{...}{additional arguments unused}
}
\value{
A logical value indicating whether the package has a BugReports field
  filled in
}
\description{
Score a package for the presence of a bug report url
}
\examples{
\dontrun{metric_score(assess_has_bug_reports_url(pkg_ref("riskmetric")))
}
}
