% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_examples.R
\name{metric_score.pkg_metric_has_examples}
\alias{metric_score.pkg_metric_has_examples}
\title{Score a package for the presence of a example or usage fields}
\usage{
\method{metric_score}{pkg_metric_has_examples}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_has_examples} packge metric object}

\item{...}{additional arguments unused}
}
\value{
\code{1} if any example or usage fields are found, otherwise \code{0}
}
\description{
Coerce the number of example or usage fields to binary indication of examples
}
\examples{
\dontrun{metric_score(assess_has_examples(pkg_ref("riskmetric")))
}
}
