% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_maintainer.R
\name{metric_score.pkg_metric_has_maintainer}
\alias{metric_score.pkg_metric_has_maintainer}
\title{Score a package for inclusion of an associated maintainer}
\usage{
\method{metric_score}{pkg_metric_has_maintainer}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_has_maintainer} packge metric object}

\item{...}{additional arguments unused}
}
\value{
\code{1} if any maintainer is provided, otherwise \code{0}
}
\description{
Coerce a list of maintainers into a numeric value indicating whether the
number of listed maintainers is greater than 0.
}
\examples{
metric_score(assess_has_maintainer(pkg_ref("riskmetric")))
}
\seealso{
Other \code{metric_score.*} functions: 
\code{\link{metric_score.pkg_metric_covr_coverage}()},
\code{\link{metric_score.pkg_metric_downloads_1yr}()},
\code{\link{metric_score.pkg_metric_export_help}()},
\code{\link{metric_score.pkg_metric_has_bug_reports_url}()},
\code{\link{metric_score.pkg_metric_has_news}()},
\code{\link{metric_score.pkg_metric_has_source_control}()},
\code{\link{metric_score.pkg_metric_has_vignettes}()},
\code{\link{metric_score.pkg_metric_has_website}()},
\code{\link{metric_score.pkg_metric_last_30_bugs_status}()},
\code{\link{metric_score.pkg_metric_license}()},
\code{\link{metric_score.pkg_metric_news_current}()},
\code{\link{metric_score.pkg_metric_r_cmd_check}()},
\code{\link{metric_score.pkg_metric_remote_checks}()}
}
\concept{\code{metric_score.*} functions}
