% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_news.R
\name{metric_score.pkg_metric_has_news}
\alias{metric_score.pkg_metric_has_news}
\title{Score a package for the presence of a NEWS file}
\usage{
\method{metric_score}{pkg_metric_has_news}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_has_news} packge metric object}

\item{...}{additional arguments unused}
}
\value{
\code{1} if any NEWS files are found, otherwise \code{0}
}
\description{
Coerce the number of news files to binary indication of valid NEWS files
}
\examples{
metric_score(assess_has_news(pkg_ref("riskmetric")))
}
