% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_source_control.R
\name{metric_score.pkg_metric_has_source_control}
\alias{metric_score.pkg_metric_has_source_control}
\title{Score a package for inclusion of an associated source control url}
\usage{
\method{metric_score}{pkg_metric_has_source_control}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_has_source_control} packge metric object}

\item{...}{additional arguments unused}
}
\value{
\code{1} if any source control url is provided, otherwise \code{0}
}
\description{
Coerce a list of source control urls into a numeric value indicating whether
the number of listed urls is greater than 0.
}
\examples{
metric_score(assess_has_source_control(pkg_ref("riskmetric")))
}
