% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_vignettes.R
\name{metric_score.pkg_metric_has_vignettes}
\alias{metric_score.pkg_metric_has_vignettes}
\title{Score a package for the presence of a Vignettes file}
\usage{
\method{metric_score}{pkg_metric_has_vignettes}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_has_vignettes} packge metric object}

\item{...}{additional arguments unused}
}
\value{
\code{1} if any Vignettes files are found, otherwise \code{0}
}
\description{
Coerce the number of vignettes files to binary indication of valid Vignettes
}
\examples{
metric_score(assess_has_vignettes(pkg_ref("riskmetric")))
}
