% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_website.R
\name{metric_score.pkg_metric_has_website}
\alias{metric_score.pkg_metric_has_website}
\title{Score a package for inclusion of an associated website url}
\usage{
\method{metric_score}{pkg_metric_has_website}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_has_website} packge metric object}

\item{...}{additional arguments unused}
}
\value{
\code{1} if any website url is provided, otherwise \code{0}
}
\description{
Coerce a list of website urls into a numeric value indicating whether the
number of listed urls is greater than 0.
}
\examples{
metric_score(assess_has_website(pkg_ref("riskmetric")))
}
