% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_last_30_bugs_status.R
\name{metric_score.pkg_metric_last_30_bugs_status}
\alias{metric_score.pkg_metric_last_30_bugs_status}
\title{Score a package for number of recently opened BugReports that are now closed}
\usage{
\method{metric_score}{pkg_metric_last_30_bugs_status}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_last_30_bugs_status} packge metric object}

\item{...}{additional arguments unused}
}
\value{
a fractional value indicating percentage of last 30 bug reports that
  are now closed
}
\description{
Score a package for number of recently opened BugReports that are now closed
}
\examples{
\dontrun{metric_score(assess_last_30_bugs_status(pkg_ref("riskmetric")))
}
}
