% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_news_current.R
\name{metric_score.pkg_metric_news_current}
\alias{metric_score.pkg_metric_news_current}
\title{Score a package for NEWS files updated to current version}
\usage{
\method{metric_score}{pkg_metric_news_current}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_news_current} packge metric object}

\item{...}{additional arguments unused}
}
\value{
\code{1} if any NEWS files are up-to-date, otherwise \code{0}
}
\description{
Coerce a logical vector of discovered up-to-date NEWS to a metric score
}
\examples{
\dontrun{metric_score(assess_news_current(pkg_ref("riskmetric")))
}
}
