% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_r_cmd_check.R
\name{metric_score.pkg_metric_r_cmd_check}
\alias{metric_score.pkg_metric_r_cmd_check}
\title{Score a package based on R CMD check results run locally}
\usage{
\method{metric_score}{pkg_metric_r_cmd_check}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_r_cmd_check} packge metric object}

\item{...}{additional arguments unused}
}
\value{
A weighted sum of errors and warnings of all tests preformed
}
\description{
The scoring function is the weighted sum of notes (0.1), errors (1) and warnings (0.25), with a maximum score of 1 (no errors, notes or warnings)
and a minimum score of 0.
Essentially, the metric will allow up to 10 notes, 1 error or 4 warnings before returning the lowest score of 0
}
\examples{
\dontrun{metric_score(assess_r_cmd_check(pkg_ref("riskmetric")))
}
}
