% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_security.R
\name{metric_score.pkg_metric_security}
\alias{metric_score.pkg_metric_security}
\title{Score a package for known security vulnerabilities in the OSS Index}
\usage{
\method{metric_score}{pkg_metric_security}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_security} packge metric object}

\item{...}{additional arguments unused}
}
\value{
A weighted sum of errors and warnings of all tests preformed
}
\description{
Coerce the count of reported vulnerabilities to a binary indicator.
}
\examples{
\dontrun{metric_score(assess_security(pkg_ref("riskmetric")))
}
}
