% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_size_codebase.R
\name{metric_score.pkg_metric_size_codebase}
\alias{metric_score.pkg_metric_size_codebase}
\title{Score a package for number of lines of code}
\usage{
\method{metric_score}{pkg_metric_size_codebase}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_size_codebase} packge metric object}

\item{...}{additional arguments unused}
}
\value{
numeric value between \code{0} (low) and \code{1} (large number of lines of code) converting the number of downloads.
}
\description{
Scores packages based on its codebase size, as determined by number of lines of code.
}
\examples{
metric_score(assess_size_codebase(pkg_ref("riskmetric")))
}
