% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_score.R
\name{metric_score}
\alias{metric_score}
\title{Score a package metric}
\usage{
metric_score(x, ...)
}
\arguments{
\item{x}{A \code{pkg_metric_*} class object to score}

\item{...}{Additional arguments unused}
}
\value{
score of a package risk metric
}
\description{
Convert a package metric into a numeric value between 0 to 1
}
