% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_score.R
\name{metric_score_condition}
\alias{metric_score_condition}
\title{Retrieve condition type raised when calculating a score}
\usage{
metric_score_condition(x, ...)
}
\arguments{
\item{x}{score of a package risk metric}

\item{...}{Additional arguments unused}
}
\value{
A classed `NA` value representing the score with condition type
}
\description{
Retrieve condition type raised when calculating a score
}
