% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_class.R
\name{new_list_of_pkg_ref}
\alias{new_list_of_pkg_ref}
\title{Create a list_list_of_pkg_ref from list of pkg_refs}
\usage{
new_list_of_pkg_ref(x)
}
\description{
Create a list_list_of_pkg_ref from list of pkg_refs
}
