% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_news.R
\name{news_from_dir}
\alias{news_from_dir}
\title{Build a list of NEWS files discovered within a given directory}
\usage{
news_from_dir(path)
}
\arguments{
\item{path}{a package directory path expected to contain NEWS files}
}
\value{
a list of parsed NEWS files
}
\description{
Build a list of NEWS files discovered within a given directory
}
