% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{not_installed_packages}
\alias{not_installed_packages}
\title{Return a vector of random uninstalled packages}
\usage{
not_installed_packages()
}
\value{
a character vector of random package names
}
\description{
Return a vector of random uninstalled packages
}
