% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_assess.R
\name{not_suggests}
\alias{not_suggests}
\title{Helper to determine if assessment depends on package in Suggests}
\usage{
not_suggests(assess_obj)
}
\arguments{
\item{assess_obj}{An assessment function object}
}
\value{
\code{TRUE} if the object does not have an attribute named
  \code{"suggests"} and the negated attribute value if present
}
\description{
Helper to determine if assessment depends on package in Suggests
}
