% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_dependencies.R
\name{parse_dcf_dependencies}
\alias{parse_dcf_dependencies}
\title{Parse DCF of description file}
\usage{
parse_dcf_dependencies(path)
}
\arguments{
\item{path}{pkg_ref path}
}
\description{
Parse DCF of description file
}
