% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_metric.R
\name{pkg_metric}
\alias{pkg_metric}
\title{A helper for structuring assessment return objects for dispatch with the
score function}
\usage{
pkg_metric(x, ..., label = NULL, class = c())
}
\description{
A helper for structuring assessment return objects for dispatch with the
score function
}
