% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_metric_error.R
\name{pkg_metric_error}
\alias{pkg_metric_error}
\title{A subclass wrapping an error with an additional parent class}
\usage{
pkg_metric_error(error)
}
\arguments{
\item{error}{an error condition object to capture}
}
\description{
A subclass wrapping an error with an additional parent class
}
