% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_metric.R
\name{pkg_metric_eval}
\alias{pkg_metric_eval}
\title{Evaluate a metric}
\usage{
pkg_metric_eval(expr, ..., class = c(), env = parent.frame())
}
\arguments{
\item{expr}{An expression to evaluate in order to calculate a
\code{pkg_metric}}

\item{...}{additional attributes to bind to the \code{pkg_metric} object}

\item{class}{a subclass to differentiate the \code{pkg_metric} object}

\item{env}{An environment in which \code{expr} is to be evaluated}
}
\value{
a \code{pkg_metric} object containing the result of \code{expr}
}
\description{
Evalute code relevant to a metric, capturing the evaluated code as well as
any messages, warnings or errors that are thrown in the process.
}
\keyword{internal}
