% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_metric_import.R
\name{pkg_metric_import}
\alias{pkg_metric_import}
\title{Import metrics}
\usage{
pkg_metric_import(file)
}
\arguments{
\item{file}{(`character`) Path to the file to import.}
}
\value{
The object in the same format and class as generated by riskmetric.
}
\description{
Import metrics
}
\seealso{
pkg_metric_export()
}
