% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_metric_na.R
\name{pkg_metric_na}
\alias{pkg_metric_na}
\title{A pkg_metric subclass for when metrics are explicitly not applicable}
\usage{
pkg_metric_na(message = NULL)
}
\description{
A pkg_metric subclass for when metrics are explicitly not applicable
}
