% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_class.R
\name{pkg_ref}
\alias{pkg_ref}
\title{Create a package reference}
\usage{
pkg_ref(x, ...)
}
\arguments{
\item{x}{character value representing a package name, path to a package
source directory or a git remote url or a package object}

\item{...}{additional arguments passed to class-specific handlers}
}
\value{
a package object
}
\description{
Create a package reference
}
