% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_bug_reports.R
\name{pkg_ref_cache.bug_reports}
\alias{pkg_ref_cache.bug_reports}
\title{Retrieve a list of BugReports metadata}
\usage{
pkg_ref_cache.bug_reports(x, ...)
}
\arguments{
\item{x}{a package reference object}

\item{...}{additional arguments used for computing cached values}
}
\description{
Retrieve a list of BugReports metadata
}
\seealso{
Other package reference cache: 
\code{\link{pkg_ref_cache.bug_reports_host}()},
\code{\link{pkg_ref_cache.bug_reports_url}()},
\code{\link{pkg_ref_cache.covr_coverage}()},
\code{\link{pkg_ref_cache.description}()},
\code{\link{pkg_ref_cache.downloads}()},
\code{\link{pkg_ref_cache.expression_coverage}()},
\code{\link{pkg_ref_cache.function_coverage}()},
\code{\link{pkg_ref_cache.help_aliases}()},
\code{\link{pkg_ref_cache.help}()},
\code{\link{pkg_ref_cache.news_urls}()},
\code{\link{pkg_ref_cache.news}()},
\code{\link{pkg_ref_cache.repo_base_url}()},
\code{\link{pkg_ref_cache.tarball_url}()},
\code{\link{pkg_ref_cache.vignettes}()},
\code{\link{pkg_ref_cache.web_html}()},
\code{\link{pkg_ref_cache.web_url}()},
\code{\link{pkg_ref_cache}()}
}
\concept{package reference cache}
