% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_bug_reports_host.R
\name{pkg_ref_cache.bug_reports_host.default}
\alias{pkg_ref_cache.bug_reports_host.default}
\title{Get the host name of a BugReports url}
\usage{
\method{pkg_ref_cache}{bug_reports_host.default}(x, ...)
}
\value{
a \code{pkg_ref} object
}
\description{
Get the host name of a BugReports url
}
\keyword{internal}
