% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_bug_reports_url.R
\name{pkg_ref_cache.bug_reports_url.pkg_source}
\alias{pkg_ref_cache.bug_reports_url.pkg_source}
\title{Get the BugReports url}
\usage{
\method{pkg_ref_cache}{bug_reports_url.pkg_source}(x, ...)
}
\value{
a \code{pkg_ref} object
}
\description{
Get the BugReports url
}
\keyword{internal}
