% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_covr_coverage.R
\name{pkg_ref_cache.covr_coverage.pkg_source}
\alias{pkg_ref_cache.covr_coverage.pkg_source}
\title{Retrieve output of covr::package_coverage}
\usage{
\method{pkg_ref_cache}{covr_coverage.pkg_source}(x, ...)
}
\value{
a \code{pkg_ref} object
}
\description{
Retrieve output of covr::package_coverage
}
\keyword{internal}
