% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_dependencies.R
\name{pkg_ref_cache.deps}
\alias{pkg_ref_cache.deps}
\title{Cache package dependencies}
\usage{
pkg_ref_cache.deps(x, ...)
}
\description{
A package's dependencies is an important indicator or the package's complexiety.
}
