% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_DESCRIPTION.R
\name{pkg_ref_cache.description}
\alias{pkg_ref_cache.description}
\title{Cache the DESCRIPTION file contents for a package reference}
\usage{
pkg_ref_cache.description(x, name, ...)
}
\arguments{
\item{x}{a package reference object}

\item{name}{the name of the field that needs to be cached}

\item{...}{additional arguments used for computing cached values}
}
\description{
Cache the DESCRIPTION file contents for a package reference
}
\seealso{
Other package reference cache: 
\code{\link{pkg_ref_cache.help_aliases}()},
\code{\link{pkg_ref_cache.help}()},
\code{\link{pkg_ref_cache.news_urls}()},
\code{\link{pkg_ref_cache.news}()},
\code{\link{pkg_ref_cache.repo_base_url}()},
\code{\link{pkg_ref_cache.tarball_url}()},
\code{\link{pkg_ref_cache.web_html}()},
\code{\link{pkg_ref_cache.web_url}()},
\code{\link{pkg_ref_cache}()}
}
\concept{package reference cache}
