% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_expr_coverage.R
\name{pkg_ref_cache.expression_coverage.pkg_source}
\alias{pkg_ref_cache.expression_coverage.pkg_source}
\title{Retrieve output of covr::package_coverage, tallied by expression}
\usage{
\method{pkg_ref_cache}{expression_coverage.pkg_source}(x, ...)
}
\value{
a \code{pkg_ref} object
}
\description{
Retrieve output of covr::package_coverage, tallied by expression
}
\keyword{internal}
