% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_expr_coverage.R
\name{pkg_ref_cache.expression_coverage}
\alias{pkg_ref_cache.expression_coverage}
\title{Retrieve output of covr::package_coverage, tallied by expression}
\usage{
pkg_ref_cache.expression_coverage(x, ...)
}
\arguments{
\item{x}{a package reference object}

\item{...}{additional arguments used for computing cached values}
}
\description{
Retrieve output of covr::package_coverage, tallied by expression
}
\seealso{
Other package reference cache: \code{\link{pkg_ref_cache.archive_release_dates}},
  \code{\link{pkg_ref_cache.bug_reports_host}},
  \code{\link{pkg_ref_cache.bug_reports_url}},
  \code{\link{pkg_ref_cache.bug_reports}},
  \code{\link{pkg_ref_cache.covr_coverage}},
  \code{\link{pkg_ref_cache.description}},
  \code{\link{pkg_ref_cache.downloads}},
  \code{\link{pkg_ref_cache.help_aliases}},
  \code{\link{pkg_ref_cache.help}},
  \code{\link{pkg_ref_cache.license}},
  \code{\link{pkg_ref_cache.maintainer}},
  \code{\link{pkg_ref_cache.news_urls}},
  \code{\link{pkg_ref_cache.news}},
  \code{\link{pkg_ref_cache.release_date}},
  \code{\link{pkg_ref_cache.repo_base_url}},
  \code{\link{pkg_ref_cache.source_control_url}},
  \code{\link{pkg_ref_cache.tarball_url}},
  \code{\link{pkg_ref_cache.vignettes}},
  \code{\link{pkg_ref_cache.web_html}},
  \code{\link{pkg_ref_cache.web_url}},
  \code{\link{pkg_ref_cache.website_urls}},
  \code{\link{pkg_ref_cache}}
}
\concept{package reference cache}
