% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_help.R
\name{pkg_ref_cache.help.pkg_install}
\alias{pkg_ref_cache.help.pkg_install}
\title{Cache a list of available help files as LaTeX objects}
\usage{
\method{pkg_ref_cache}{help.pkg_install}(x, name, ...)
}
\value{
a \code{pkg_ref} object
}
\description{
Cache a list of available help files as LaTeX objects
}
\keyword{internal}
