% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_news.R
\name{pkg_ref_cache.news.pkg_remote}
\alias{pkg_ref_cache.news.pkg_remote}
\title{Cache a list of NEWS files from a package reference}
\usage{
\method{pkg_ref_cache.news}{pkg_remote}(x, name, ...)
}
\value{
a \code{pkg_ref} object
}
\description{
Cache a list of NEWS files from a package reference
}
\keyword{internal}
