% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_news.R
\name{pkg_ref_cache.news}
\alias{pkg_ref_cache.news}
\title{Cache a list of NEWS files from a package reference}
\usage{
pkg_ref_cache.news(x, name, ...)
}
\description{
Cache a list of NEWS files from a package reference
}
\seealso{
Other package reference cache: \code{\link{pkg_ref_cache.description}},
  \code{\link{pkg_ref_cache.help_aliases}},
  \code{\link{pkg_ref_cache.help}},
  \code{\link{pkg_ref_cache.news_urls}},
  \code{\link{pkg_ref_cache.repo_base_url}},
  \code{\link{pkg_ref_cache.tarball_url}},
  \code{\link{pkg_ref_cache.web_html}},
  \code{\link{pkg_ref_cache.web_url}},
  \code{\link{pkg_ref_cache}}
}
\concept{package reference cache}
