% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_news_urls.R
\name{pkg_ref_cache.news_urls}
\alias{pkg_ref_cache.news_urls}
\title{Cache appropriate urls for NEWS files}
\usage{
pkg_ref_cache.news_urls(x, name, ...)
}
\description{
Cache appropriate urls for NEWS files
}
\seealso{
Other package reference cache: \code{\link{pkg_ref_cache.description}},
  \code{\link{pkg_ref_cache.help_aliases}},
  \code{\link{pkg_ref_cache.help}},
  \code{\link{pkg_ref_cache.news}},
  \code{\link{pkg_ref_cache.repo_base_url}},
  \code{\link{pkg_ref_cache.tarball_url}},
  \code{\link{pkg_ref_cache.web_html}},
  \code{\link{pkg_ref_cache.web_url}},
  \code{\link{pkg_ref_cache}}
}
\concept{package reference cache}
