% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_r_cmd_check.R
\name{pkg_ref_cache.r_cmd_check.pkg_source}
\alias{pkg_ref_cache.r_cmd_check.pkg_source}
\title{Run R CMD check and capture the results}
\usage{
\method{pkg_ref_cache}{r_cmd_check.pkg_source}(x, ...)
}
\value{
a \code{pkg_ref} object
}
\description{
Run R CMD check and capture the results
}
