% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_r_cmd_check.R
\name{pkg_ref_cache.r_cmd_check}
\alias{pkg_ref_cache.r_cmd_check}
\title{Run R CMD check and capture the results}
\description{
Run R CMD check and capture the results
}
\seealso{
Other package reference cache: 
\code{\link{pkg_ref_cache}()}
}
\concept{package reference cache}
\keyword{internal}
