% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache.R
\name{pkg_ref_cache}
\alias{pkg_ref_cache}
\title{S3 function to extend a package reference with new fields}
\usage{
pkg_ref_cache(x, name, ..., .class = as.character(name))
}
\arguments{
\item{x}{a package reference object}

\item{name}{the name of the field that needs to be cached}

\item{...}{additional arguments used for computing cached values}

\item{.class}{a class name to use for S3 dispatch, defaulting to the name as
a character value}
}
\value{
a value to assign to the new field in the package reference object
  environment
}
\description{
S3 function to extend a package reference with new fields
}
\seealso{
Other package reference cache: \code{\link{pkg_ref_cache.description}},
  \code{\link{pkg_ref_cache.help_aliases}},
  \code{\link{pkg_ref_cache.help}},
  \code{\link{pkg_ref_cache.news_urls}},
  \code{\link{pkg_ref_cache.news}},
  \code{\link{pkg_ref_cache.repo_base_url}},
  \code{\link{pkg_ref_cache.tarball_url}},
  \code{\link{pkg_ref_cache.web_html}},
  \code{\link{pkg_ref_cache.web_url}}
}
\concept{package reference cache}
