% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_security.R
\name{pkg_ref_cache.security.default}
\alias{pkg_ref_cache.security.default}
\title{Check OSS Index lists any vulnerabilities for the package and it's
dependencies}
\usage{
\method{pkg_ref_cache}{security.default}(x, ...)
}
\arguments{
\item{x}{a package reference object}

\item{...}{additional arguments used for computing cached values}
}
\value{
a \code{pkg_ref} object
}
\description{
Check OSS Index lists any vulnerabilities for the package and it's
dependencies
}
