% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_security.R
\name{pkg_ref_cache.security}
\alias{pkg_ref_cache.security}
\title{Cache OSS Index results}
\usage{
pkg_ref_cache.security(x, ...)
}
\value{
a \code{pkg_ref} object
}
\description{
Cache OSS Index results
}
\concept{package reference cache}
\keyword{internal}
