% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_vignettes.R
\name{pkg_ref_cache.vignettes}
\alias{pkg_ref_cache.vignettes}
\title{Cache a List of Vignettes Files from a Package Reference}
\usage{
pkg_ref_cache.vignettes(x, name, ...)
}
\description{
Cache a List of Vignettes Files from a Package Reference
}
\concept{package reference cache}
\keyword{internal}
