% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_class.R
\docType{data}
\name{pkg_ref_class_hierarchy}
\alias{pkg_ref_class_hierarchy}
\title{The `pkg_ref` subclass hierarchy, used for pkg_ref object creation with a
specified subclass}
\format{
An object of class \code{list} of length 1.
}
\usage{
pkg_ref_class_hierarchy
}
\description{
The `pkg_ref` subclass hierarchy, used for pkg_ref object creation with a
specified subclass
}
\keyword{datasets}
