% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_reference.R
\name{pkg_reference}
\alias{pkg_reference}
\title{Convert into a package object}
\usage{
pkg_reference(x, ...)
}
\arguments{
\item{x}{character value representing a package name, path to a package
source directory or a git remote url or a package object}

\item{...}{additional arguments passed to class-specific handlers}
}
\value{
a package object
}
\description{
Convert into a package object
}
