% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_risk.R
\name{pkg_risk}
\alias{pkg_risk}
\title{Calculate a set of package risk scores}
\usage{
pkg_risk(
  x,
  source = c(NA, "pkg_install", "pkg_source", "pkg_missing", "pkg_cran_remote",
    "pkg_bioc_remote", "pkg_git_remote"),
  lib.loc = .libPaths(),
  assessments = all_assessments(),
  repos = getOption("repos", "https://cran.rstudio.com")
)
}
\arguments{
\item{x}{A singular \code{character} value, \code{character vector} or
\code{list} of \code{character} values of package names or source code
directory paths.}

\item{source}{Optional argument that will overright what riskmetric does when
determining a package source. By default it will find sources for package
info based on the riskmetric subclass hierarchy and availability on your
system.}

\item{lib.loc}{The path to the R library directory of the installed package.}

\item{assessments}{A list of assessment functions to apply to each package
reference. By default, a list of all exported assess_* functions from the
riskmetric package.}

\item{repos}{URL of CRAN repository to pull package metadata. Defaults to
global option set on your system or the RStudio CRAN URL if not set.}
}
\value{
A numeric value if a single \code{pkg_metric} is provided, or a
  \code{\link[tibble]{tibble}} with \code{pkg_metric} objects scored and
  returned as numeric values when a \code{\link[tibble]{tibble}} is provided.
}
\description{
Calculate a set of package risk scores
}
\examples{
\dontrun{

# scoring a single package
pkg_risk("riskmetric")

# scoring many package as a tibble
pkg_risk(c("riskmetric", "dplyr", "abc"))

# score packages from a specific package source
pkg_risk(c("riskmetric", "dplyr", "abc"), source = "pkg_cran_remote")

}

}
