% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_risk.R
\name{pkg_risk_cran}
\alias{pkg_risk_cran}
\title{Calculate a set of package risk scores on CRAN}
\usage{
pkg_risk_cran(x, source = "pkg_cran_remote", ...)
}
\arguments{
\item{x}{A singular \code{character} value, \code{character vector} or
\code{list} of \code{character} values of package names or source code
directory paths.}

\item{source}{Optional argument that will overright what riskmetric does when
determining a package source. By default it will find sources for package
info based on the riskmetric subclass hierarchy and availability on your
system.}

\item{...}{Additional arguments passed to `pkg_risk()`}
}
\description{
Calculate a set of package risk scores on CRAN
}
\examples{
\dontrun{
# score packages from a specific package source
pkg_risk_cran(c("riskmetric", "dplyr", "abc"))
}

}
