% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.expr_output}
\alias{print.expr_output}
\title{Handle pretty printing of expression output}
\usage{
\method{print}{expr_output}(x, ...)
}
\arguments{
\item{x}{expr_output to print}

\item{...}{additional arguments unused}
}
\description{
Handle pretty printing of expression output
}
