% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.with_eval_recording}
\alias{print.with_eval_recording}
\title{Handle pretty printing of expression output}
\usage{
\method{print}{with_eval_recording}(x, playback = FALSE, cr = TRUE, ..., sleep = 0)
}
\arguments{
\item{x}{expr_output to print}

\item{playback}{a \code{logical} indicating whether evaluation output
should be played back (\code{FALSE}), or whether the result value should
be printed as is (\code{TRUE}, the default)}

\item{cr}{a \code{logical} indicating whether carriage returns should be
printed, possibly overwriting characters in the output.}

\item{...}{additional arguments unused}

\item{sleep}{an \code{numeric} indicating a time to sleep between printing
each line to console. This can be helpful if the original output overwrites
valuable information in the log that is eventually overwritten and you
would like to watch it play out as it was formatted.}
}
\value{
a print message
}
\description{
Handle pretty printing of expression output
}
\keyword{internal}
