% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoteReference.R
\name{RemoteReference}
\alias{RemoteReference}
\title{Remote reference to package sources}
\usage{
RemoteReference(x, type, version = "latest", repos = list(CRAN =
  "https://cloud.r-project.org"))
}
\arguments{
\item{x}{package/repo name}

\item{type}{"cran" or "github"}

\item{version}{version tag to use for CRAN e.g. 0.7.1, release tag for github}

\item{repos}{list with single CRAN-like repository to query sources from}
}
\description{
Remote reference to package sources
}
