% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_dependencies.R
\name{remove_base_packages}
\alias{remove_base_packages}
\title{Helper function to remove base and recommended packages}
\usage{
remove_base_packages(df)
}
\arguments{
\item{df}{Data frame of dependencies of a package.}
}
\description{
Helper function to remove base and recommended packages
}
