% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache__behaviors.R
\name{require_cache_behaviors}
\alias{require_cache_behaviors}
\title{Stop if a function requires disabled behaviors}
\usage{
require_cache_behaviors(behaviors)
}
\arguments{
\item{per_package_request}{declared when caching a value requires a web request for each package individually, which can be time intensive and an irresponsible use of generously hosted public R package repositories. It is recommended that this behavior is disabled for metric assessments of large numbers of packages or assessments triggered via automated scripts. Automatically disabled by default for non-interactive use.}
}
\description{
Stop if a function requires disabled behaviors
}
