% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskmetric-package.R
\docType{package}
\name{riskmetric}
\alias{riskmetric-package}
\alias{riskmetric}
\title{riskmetric}
\description{
Facilities for assessing R packages against a number of metrics to help
quantify their robustness.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmar.github.io/riskmetric/}
  \item \url{https://github.com/pharmaR/riskmetric}
  \item Report bugs at \url{https://github.com/pharmaR/riskmetric/issues}
}

}
\author{
\strong{Maintainer}: Eli Miller \email{eli.miller@atorusresearch.com}

Authors:
\itemize{
  \item R Validation Hub \email{psi.aims.r.validation@gmail.com}
  \item Doug Kelkhoff \email{doug.kelkhoff@gmail.com}
  \item Marly Gotti
  \item Kevin K
  \item Yilong Zhang
  \item Eric Milliman
  \item Juliane Manitz
}

Other contributors:
\itemize{
  \item Mark Padgham [contributor]
  \item PSI special interest group Application and Implementation of Methodologies in Statistics [copyright holder]
}

}
\keyword{internal}
