% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_assess.R
\name{roxygen_assess_family_catalog}
\alias{roxygen_assess_family_catalog}
\title{Helper for creating a roxygen itemized list for assess_* functions}
\usage{
roxygen_assess_family_catalog()
}
\value{
roxygen section template for assess family function catalog
}
\description{
Helper for creating a roxygen itemized list for assess_* functions
}
\examples{
\dontrun{
  #' @eval assess_family_catalog_roxygen()
}

}
