% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_downloads.r
\name{score.pkg_metric_downloads_1yr}
\alias{score.pkg_metric_downloads_1yr}
\title{Score a package for the number of downloads in the past year}
\usage{
\method{score}{pkg_metric_downloads_1yr}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_downloads_1yr} packge metric object}

\item{...}{additional arguments unused}
}
\value{
numeric value between \code{0} (low) and \code{1} (high download volume) converting the number of downloads.
}
\description{
Convert the number of downloads \code{x} in the past year into a validation score [0,1] 
\deqn{ 1 - \frac{150,000}{X + 150,000} }
}
\examples{
score(assess_downloads_1yr(pkg_ref("riskmetric")))
}
\seealso{
Other \code{score.*} functions: \code{\link{score.pkg_metric_export_help}},
  \code{\link{score.pkg_metric_has_news}},
  \code{\link{score.pkg_metric_news_current}},
  \code{\link{score}}
}
\concept{\code{score.*} functions}
