% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_export_help.R
\name{score.pkg_metric_export_help}
\alias{score.pkg_metric_export_help}
\title{Score a package for availability of documentation for exported values}
\usage{
\method{score}{pkg_metric_export_help}(x)
}
\arguments{
\item{x}{a \code{pkg_metric_export_help} packge metric object}

\item{...}{additional arguments unused}
}
\value{
\code{1} if any NEWS files are found, otherwise \code{0}
}
\description{
Coerce a logical vector indicating availability of export documentation
}
\examples{
score(assess_export_help(pkg_ref("riskmetric")))
}
\seealso{
Other \code{score.*} functions: \code{\link{score.pkg_metric_has_news}},
  \code{\link{score.pkg_metric_news_current}},
  \code{\link{score}}
}
\concept{\code{score.*} functions}
