% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_news.R
\name{score.pkg_metric_has_news}
\alias{score.pkg_metric_has_news}
\title{Score a package for the presence of a NEWS file}
\usage{
\method{score}{pkg_metric_has_news}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_has_news} packge metric object}

\item{...}{additional arguments unused}
}
\value{
\code{1} if any NEWS files are found, otherwise \code{0}
}
\description{
Coerce the number of news files to binary indication of a valid NEWS
}
\examples{
score(assess_has_news(pkg_ref("riskmetric")))
}
\seealso{
Other \code{score.*} functions: \code{\link{score.pkg_metric_export_help}},
  \code{\link{score.pkg_metric_news_current}},
  \code{\link{score}}
}
\concept{\code{score.*} functions}
