% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_vignettes.R
\name{score.pkg_metric_has_vignettes}
\alias{score.pkg_metric_has_vignettes}
\title{Score a package for the presence of a Vignettes file}
\usage{
\method{score}{pkg_metric_has_vignettes}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_has_vignettes} packge metric object}

\item{...}{additional arguments unused}
}
\value{
\code{1} if any Vignettes files are found, otherwise \code{0}
}
\description{
Coerce the number of vignettes files to binary indication of valid Vignettes
}
\examples{
score(assess_has_vignettes(pkg_ref("riskmetric")))
}
\seealso{
Other \code{score.*} functions: 
\code{\link{score.pkg_metric_covr_coverage}()},
\code{\link{score.pkg_metric_downloads_1yr}()},
\code{\link{score.pkg_metric_export_help}()},
\code{\link{score.pkg_metric_has_bug_reports_url}()},
\code{\link{score.pkg_metric_has_news}()},
\code{\link{score.pkg_metric_last_30_bugs_status}()},
\code{\link{score.pkg_metric_license}()},
\code{\link{score.pkg_metric_news_current}()},
\code{\link{score}()}
}
\concept{\code{score.*} functions}
