% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_last_30_bugs_status.R
\name{score.pkg_metric_last_30_bugs_status}
\alias{score.pkg_metric_last_30_bugs_status}
\title{Score a package for number of recently opened BugReports that are now closed}
\usage{
\method{score}{pkg_metric_last_30_bugs_status}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_last_30_bugs_status} packge metric object}

\item{...}{additional arguments unused}
}
\value{
a fractional value indicating percentage of last 30 bug reports that
  are now closed
}
\description{
Score a package for number of recently opened BugReports that are now closed
}
\examples{
\dontrun{score(assess_last_30_bugs_status(pkg_ref("riskmetric")))
}
}
\seealso{
Other \code{score.*} functions: 
\code{\link{score.pkg_metric_covr_coverage}()},
\code{\link{score.pkg_metric_downloads_1yr}()},
\code{\link{score.pkg_metric_export_help}()},
\code{\link{score.pkg_metric_has_news}()},
\code{\link{score.pkg_metric_has_vignettes}()},
\code{\link{score.pkg_metric_license}()},
\code{\link{score.pkg_metric_news_current}()},
\code{\link{score}()}
}
\concept{\code{score.*} functions}
