% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_license.R
\name{score.pkg_metric_license}
\alias{score.pkg_metric_license}
\title{Score a package for acceptable license}
\usage{
\method{score}{pkg_metric_license}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_license} packge metric object}

\item{...}{additional arguments unused}
}
\value{

}
\description{
Maps a license string to a score
}
\examples{
score(assess_license(pkg_ref("riskmetric")))
}
\seealso{
Other \code{score.*} functions: 
\code{\link{score.pkg_metric_export_help}()},
\code{\link{score.pkg_metric_has_news}()},
\code{\link{score.pkg_metric_news_current}()},
\code{\link{score}()}
}
\concept{\code{score.*} functions}
