% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_news_current.R
\name{score.pkg_metric_news_current}
\alias{score.pkg_metric_news_current}
\title{Score a package for NEWS files updated to current version}
\usage{
\method{score}{pkg_metric_news_current}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_news_current} packge metric object}

\item{...}{additional arguments unused}
}
\value{
\code{1} if any NEWS files are up-to-date, otherwise \code{0}
}
\description{
Coerce a logical vector of discovered up-to-date NEWS to a metric score
}
\examples{
score(assess_news_current(pkg_ref("riskmetric")))
}
\seealso{
<<<<<<< HEAD
Other \code{score.*} functions: \code{\link{score.pkg_metric_export_help}},
  \code{\link{score.pkg_metric_has_news}},
  \code{\link{score.pkg_metric_last_30_bugs_status}},
  \code{\link{score}}
=======
Other \code{score.*} functions: 
\code{\link{score.pkg_metric_downloads_1yr}()},
\code{\link{score.pkg_metric_export_help}()},
\code{\link{score.pkg_metric_has_news}()},
\code{\link{score.pkg_metric_license}()},
\code{\link{score}()}
>>>>>>> master
}
\concept{\code{score.*} functions}
