% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_metric_has_news.R, R/score.R
\name{score.pkg_metric_has_news}
\alias{score.pkg_metric_has_news}
\alias{score}
\alias{score.default}
\alias{score.tbl_df}
\title{Score a package assement, collapsing results into a single numeric}
\usage{
\method{score}{pkg_metric_has_news}(x, ...)

score(x, ..., envir = parent.frame())

\method{score}{default}(x, ..., envir = parent.frame())

\method{score}{tbl_df}(x, ..., envir = parent.frame())
}
\description{
Score a package assement, collapsing results into a single numeric
}
