% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_score.R
\name{score_error_default}
\alias{score_error_default}
\title{Default score error handling, emitting a warning and returning 0}
\usage{
score_error_default(x, ...)
}
\arguments{
\item{x}{A \code{pkg_metric_*} class object to score}

\item{...}{Additional arguments unused}
}
\description{
Default score error handling, emitting a warning and returning 0
}
\seealso{
Other \code{score_error_*} functions: \code{\link{score_error_NA}},
  \code{\link{score_error_zero}}
}
\concept{\code{score_error_*} functions}
