% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_error_NA}
\alias{score_error_NA}
\title{Score error handler to silently return NA}
\usage{
score_error_NA(...)
}
\arguments{
\item{...}{additional arguments unused}
}
\description{
Score error handler to silently return NA
}
\seealso{
Other \code{score_error_*} functions: 
\code{\link{score_error_default}()},
\code{\link{score_error_zero}()}
}
\concept{\code{score_error_*} functions}
