% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_metric_weight.R
\name{set_metric_weight}
\alias{set_metric_weight}
\title{Function to conviently set the weights for a subset of package metrics}
\usage{
set_metric_weight(data, weight)
}
\arguments{
\item{data}{`pkg_assess` object}

\item{weight}{named vector of assessment(s) and their weight(s)}
}
\value{
A named vector of package metrics and their weights
}
\description{
Sometimes you want to change one or a few metric weights while keeping the rest the same,
instead of having to define all weights this function will let you set the weight of
specific metrics while returning the default weight for the rest of the metrics not named.
}
