% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_class_extract.R
\name{[[.pkg_ref}
\alias{[[.pkg_ref}
\title{Lazily instantiated, immutable metadata access}
\usage{
\method{[[}{pkg_ref}(x, name, ...)
}
\arguments{
\item{x}{pkg_ref object to extract metadata from}

\item{name}{name of metadata field to extract}

\item{...}{additional arguments used to extract from internal environment}
}
\description{
If errors are thrown upon instantiation, they are saved and rethrown any time
the value is attempted to be accessed. These then propegate through
assessment and scoring functions to affect any downstream metrics.
}
\keyword{internal}
