% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_suggests.R
\name{suggests_install_helper}
\alias{suggests_install_helper}
\title{Helper function to guide user through install of a package}
\usage{
suggests_install_helper(pkg_name, calling_fn)
}
\description{
Helper function to guide user through install of a package
}
\keyword{internal}
