% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_risk.R
\name{summarize_risk}
\alias{summarize_risk}
\title{Summarize a default set of assessments into a single risk score}
\usage{
summarize_risk(data, weights = NULL)
}
\arguments{
\item{data}{a \code{\link[tibble]{tibble}} of scored assessments whose column
names match those provided by riskmetric's \code{\link{assess}} function.}

\item{weights}{a set of numeric weights to give to each score column when
calculating risk}
}
\value{
a numeric vector of risk scores
}
\description{
This function serves as an example for how a risk score might be derived.
Assuming all assessments provided by \code{riskmetric} are available in a
dataset, this function can be used to calculate a vector of risks.
}
