% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{suppressMatching}
\alias{suppressMatching}
\title{Suppress messages and warnings based on one or more regex matches}
\usage{
suppressMatching(expr, messages = NULL, warnings = NULL, ...)
}
\arguments{
\item{expr}{An expression to evaluate}

\item{messages, warnings}{A vector of regular expressions that, when matched
against the respective condition message, should suppress that condition.}

\item{...}{Additional arguments passed to \code{grepl}}
}
\description{
Suppress messages and warnings based on one or more regex matches
}
