% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{suppressMatchingConditions}
\alias{suppressMatchingConditions}
\title{Suppress messages and warnings based on one or more regex matches}
\usage{
suppressMatchingConditions(expr, ..., .opts = list(), .envir = parent.frame())
}
\arguments{
\item{expr}{An expression to evaluate}

\item{...}{Named parameters, where the name indicates the class of conditions
to capture and the value is a vector of regular expressions that, when
matched against the respective condition message, should suppress that
condition.}

\item{.opts}{A named list of arguments to pass to \code{grepl}}

\item{.envir}{The environment in which \code{expr} is to be evaluated}
}
\value{
a message printed on console
}
\description{
Suppress messages and warnings based on one or more regex matches
}
\keyword{internal}
