% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{tbl_coverage.PackageDirectory}
\alias{tbl_coverage.PackageDirectory}
\title{[WIP] not working - coverage is always empty}
\usage{
\method{tbl_coverage}{PackageDirectory}(pkg_ref, ...)
}
\description{
[WIP] not working - coverage is always empty
}
