% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{use_assessments_column_names}
\alias{use_assessments_column_names}
\title{reassign assignment list names with column_name attribute if available}
\usage{
use_assessments_column_names(x)
}
\description{
reassign assignment list names with column_name attribute if available
}
