% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_suggests.R
\name{validate_suggests_install}
\alias{validate_suggests_install}
\title{Helper function to prompt to install a package if it is not present}
\usage{
validate_suggests_install(pkg_name, calling_fn = deparse(match.call()[[1]]))
}
\description{
Helper function to prompt to install a package if it is not present
}
\keyword{internal}
