% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_class.R
\name{verify_pkg_source}
\alias{verify_pkg_source}
\title{Verify a pkg_source when one is manually specified by the user}
\usage{
verify_pkg_source(x, source, repos)
}
\description{
Verify a pkg_source when one is manually specified by the user
}
\keyword{internal}
