% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_vignettes.R
\name{vignettes_from_dir}
\alias{vignettes_from_dir}
\title{Build a List of Vignettes Files Discovered Within a Given Directory}
\usage{
vignettes_from_dir(path)
}
\arguments{
\item{path}{a package directory path expected to contain Vignettes files}
}
\value{
a vector of parsed Vignettes files
}
\description{
Build a List of Vignettes Files Discovered Within a Given Directory
}
\keyword{internal}
