% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_vignettes.R
\name{vignettes_from_html}
\alias{vignettes_from_html}
\title{Build a List of Vignettes Files Discovered Within a Package Website}
\usage{
vignettes_from_html(x)
}
\arguments{
\item{x}{a \code{pkg_ref} object}
}
\value{
a vector of Vignettes files
}
\description{
Build a List of Vignettes Files Discovered Within a Package Website
}
