% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_unclassed_to}
\alias{with_unclassed_to}
\title{Evaluate an expression after first removing a range of S3 classes}
\usage{
with_unclassed_to(x, .class = 1:length(class(x)), expr,
  envir = parent.frame())
}
\description{
Evaluate an expression after first removing a range of S3 classes
}
