% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_assess.R
\name{all_assessments}
\alias{all_assessments}
\title{A default list of assessments to perform for each package}
\usage{
all_assessments(include_suggests = FALSE)
}
\arguments{
\item{include_suggests}{Logical indicating if assessments requiring
dependency packages listed in \code{Suggests} be included in the list of
assessments?}
}
\value{
a list of assess_* functions exported from riskmetric
}
\description{
A default list of assessments to perform for each package
}
