% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_class_extract.R
\name{allow_mutation}
\alias{allow_mutation}
\title{a wrapper to assert that a pkg_ref has been permitted to do an additional
mutation, used to handle recursive initialization of cached fields}
\usage{
allow_mutation(x, expr, envir = parent.frame())
}
\description{
a wrapper to assert that a pkg_ref has been permitted to do an additional
mutation, used to handle recursive initialization of cached fields
}
