% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref.R
\name{as.pkg_ref}
\alias{as.pkg_ref}
\title{Convert into a package object}
\usage{
as.pkg_ref(x, ...)
}
\arguments{
\item{x}{character value representing a package name, path to a package
source directory or a git remote url or a package object}
}
\value{
a package object
}
\description{
Convert into a package object
}
