% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_class.R
\name{as_list_of_pkg_ref}
\alias{as_list_of_pkg_ref}
\title{Coerce a list or pkg_ref to a list_list_of_pkg_ref}
\usage{
as_list_of_pkg_ref(x)
}
\arguments{
\item{x}{list of pkg_refs to be coerced to a `list_of_pkg_ref`}
}
\description{
Coerce a list or pkg_ref to a list_list_of_pkg_ref
}
