% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_metric_condition.R
\name{as_pkg_metric_condition}
\alias{as_pkg_metric_condition}
\title{A pkg_metric subclass for general metric evaluation conditions}
\usage{
as_pkg_metric_condition(x, ..., subclass = c())
}
\arguments{
\item{x}{an object to wrap in a \code{pkg_metric_condition} class}

\item{...}{additional arguments added as attributes to object \code{x}}

\item{subclass}{an optional subclass of \code{pkg_metric_condition} to
include}
}
\description{
A pkg_metric subclass for general metric evaluation conditions
}
\keyword{internal}
