% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_metric_error.R
\name{as_pkg_metric_error}
\alias{as_pkg_metric_error}
\title{A subclass wrapping an error with an additional parent class}
\usage{
as_pkg_metric_error(error)
}
\arguments{
\item{error}{an error condition object to capture}
}
\value{
an error condition object after wrap \code{pkg_metric_error} class.
}
\description{
A subclass wrapping an error with an additional parent class
}
\keyword{internal}
