% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_covr_coverage.R
\name{assess_covr_coverage}
\alias{assess_covr_coverage}
\title{Assess a package code coverage using the `covr` package}
\usage{
assess_covr_coverage(x, ...)
}
\description{
Assess a package code coverage using the `covr` package
}
