% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_covr_coverage.R
\name{assess_covr_coverage}
\alias{assess_covr_coverage}
\title{Assess a package for an up-to-date NEWS file}
\usage{
assess_covr_coverage(x, ...)
}
\arguments{
\item{x}{a \code{pkg_ref} package reference object}

\item{...}{additional arguments passed on to S3 methods, rarely used}
}
\value{
a \code{pkg_metric} containing a list containing fields 'filecoverage' and 'totalcoverage' containing a named numeric vector of file unit test coverage and a singular numeric value representing overall test coverage respectively.
}
\description{
Assess a package for an up-to-date NEWS file
}
\examples{
assess_news_current(pkg_ref("riskmetric"))
}
\seealso{
\code{\link{score.pkg_metric_news_current}}

Other \code{assess_*} functions: 
\code{\link{assess_downloads_1yr}()},
\code{\link{assess_export_help}()},
\code{\link{assess_has_news}()},
\code{\link{assess_has_vignettes}()},
\code{\link{assess_last_30_bugs_status}()},
\code{\link{assess_license}()},
\code{\link{assess_news_current}()},
\code{\link{assess}()}
}
\concept{\code{assess_*} functions}
