% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_downloads.R
\name{assess_downloads_1yr}
\alias{assess_downloads_1yr}
\title{Assess a package for the number of downloads in the past year}
\usage{
assess_downloads_1yr(x, ...)
}
\arguments{
\item{x}{a \code{pkg_ref} package reference object}

\item{...}{additional arguments passed on to S3 methods, rarely used}
}
\value{
a \code{pkg_metric} containing a numeric value between [0,1] indicating the volume of downloads
}
\description{
Assess a package for the number of downloads in the past year
}
\details{
The more times a package has been downloaded the more extensive the user testing and the greater chance there is of someone finding a bug and logging it.
}
\examples{
assess_downloads_1yr(pkg_ref("riskmetric"))
}
\seealso{
\code{\link{metric_score.pkg_metric_downloads_1yr}}

Other \code{assess_*} functions: 
\code{\link{assess_covr_coverage}()},
\code{\link{assess_export_help}()},
\code{\link{assess_has_bug_reports_url}()},
\code{\link{assess_has_maintainer}()},
\code{\link{assess_has_news}()},
\code{\link{assess_has_source_control}()},
\code{\link{assess_has_vignettes}()},
\code{\link{assess_has_website}()},
\code{\link{assess_last_30_bugs_status}()},
\code{\link{assess_license}()},
\code{\link{assess_news_current}()},
\code{\link{assess_remote_checks}()},
\code{\link{pkg_assess}()}
}
\concept{\code{assess_*} functions}
