% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_export_help.R
\name{assess_export_help}
\alias{assess_export_help}
\title{Assess a package for availability of documentation for exported values}
\usage{
assess_export_help(x, ...)
}
\description{
Assess a package for availability of documentation for exported values
}
