% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_export_help.R
\name{assess_export_help}
\alias{assess_export_help}
\title{Assess a package for availability of documentation for exported values}
\usage{
assess_export_help(x, ...)
}
\arguments{
\item{x}{a \code{pkg_ref} package reference object}

\item{...}{additional arguments passed on to S3 methods, rarely used}
}
\value{
a \code{pkg_metric} containing a logical vector indicating existence of documentation for each namespace export
}
\description{
Assess a package for availability of documentation for exported values
}
\examples{
assess_export_help(pkg_ref("riskmetric"))
}
\seealso{
\code{\link{score.pkg_metric_export_help}}

<<<<<<< HEAD
Other \code{assess_*} functions: \code{\link{assess_has_news}},
  \code{\link{assess_last_30_bugs_status}},
  \code{\link{assess_news_current}}
=======
Other \code{assess_*} functions: 
\code{\link{assess_downloads_1yr}()},
\code{\link{assess_has_news}()},
\code{\link{assess_license}()},
\code{\link{assess_news_current}()},
\code{\link{assess}()}
>>>>>>> master
}
\concept{\code{assess_*} functions}
