% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_exported_namespace.R
\name{assess_exported_namespace}
\alias{assess_exported_namespace}
\title{Assess a package's results from running R CMD check}
\usage{
assess_exported_namespace(x, ...)
}
\description{
Assess a package's results from running R CMD check
}
