% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_maintainer.R
\name{assess_has_maintainer}
\alias{assess_has_maintainer}
\title{Assess a package for an associated maintainer}
\usage{
assess_has_maintainer(x, ...)
}
\arguments{
\item{x}{a \code{pkg_ref} package reference object}

\item{...}{additional arguments passed on to S3 methods, rarely used}
}
\value{
a \code{pkg_metric} containing a character vector of maintainers associated with the package
}
\description{
Assess a package for an associated maintainer
}
\examples{
assess_has_maintainer(pkg_ref("riskmetric"))
}
\seealso{
\code{\link{metric_score.pkg_metric_has_maintainer}}
}
