% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_news.R
\name{assess_has_news}
\alias{assess_has_news}
\title{Assess a package for the presence of a NEWS file}
\usage{
assess_has_news(x, ...)
}
\arguments{
\item{x}{a \code{pkg_ref} package reference object}

\item{...}{additional arguments passed on to S3 methods, rarely used}
}
\value{
a \code{pkg_metric} containing an integer value indicating the number of discovered NEWS files
}
\description{
Assess a package for the presence of a NEWS file
}
\examples{
assess_has_news(pkg_ref("riskmetric"))
}
\seealso{
\code{\link{score.pkg_metric_has_news}}

Other \code{assess_*} functions: \code{\link{assess_export_help}},
  \code{\link{assess_news_current}}
}
\concept{\code{assess_*} functions}
