% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_metric_has_news.R
\name{assess_has_news}
\alias{assess_has_news}
\alias{assess_has_news.pkg_remote}
\alias{assess_has_news.pkg_install}
\alias{assess_has_news.pkg_source}
\title{Assess a package for the presence of a NEWS file}
\usage{
assess_has_news(x, ...)

\method{assess_has_news}{pkg_remote}(x, ...)

\method{assess_has_news}{pkg_install}(x, ...)

\method{assess_has_news}{pkg_source}(x, ...)
}
\arguments{
\item{x}{a packge reference object}

\item{...}{additional arguments unused}
}
\value{
a \code{pkg_metric} object
}
\description{
Assess a package for the presence of a NEWS file
}
