% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_source_control.R
\name{assess_has_source_control}
\alias{assess_has_source_control}
\title{Assess a package for an associated source control url}
\usage{
assess_has_source_control(x, ...)
}
\arguments{
\item{x}{a \code{pkg_ref} package reference object}

\item{...}{additional arguments passed on to S3 methods, rarely used}
}
\value{
a \code{pkg_metric} containing a character vector of source control urls associated with the package
}
\description{
Assess a package for an associated source control url
}
\examples{
assess_has_source_control(pkg_ref("riskmetric"))
}
\seealso{
\code{\link{metric_score.pkg_metric_has_source_control}}
}
