% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_last_30_bugs_status.R
\name{assess_last_30_bugs_status}
\alias{assess_last_30_bugs_status}
\title{Assess how many recent BugReports have been closed}
\usage{
assess_last_30_bugs_status(x, ...)
}
\arguments{
\item{x}{a \code{pkg_ref} package reference object}

\item{...}{additional arguments passed on to S3 methods, rarely used}
}
\value{
a \code{pkg_metric} containing a logical vector indicating whether a recent BugReport was closed
}
\description{
Assess how many recent BugReports have been closed
}
\examples{
\dontrun{
assess_last_30_bugs_status(pkg_ref("riskmetric"))
}
}
\seealso{
\code{\link{score.pkg_metric_last_30_bugs_status}}

Other \code{assess_*} functions: 
\code{\link{assess_downloads_1yr}()},
\code{\link{assess_export_help}()},
\code{\link{assess_has_news}()},
\code{\link{assess_has_vignettes}()},
\code{\link{assess_license}()},
\code{\link{assess_news_current}()},
\code{\link{assess}()}
}
\concept{\code{assess_*} functions}
