% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_license.R
\name{assess_license}
\alias{assess_license}
\title{Assess a package for an up-to-date NEWS file}
\usage{
assess_license(x, ...)
}
\arguments{
\item{x}{a \code{pkg_ref} package reference object}

\item{...}{additional arguments passed on to S3 methods, rarely used}
}
\value{
a \code{pkg_metric} containing a string indicating the license under which the package is released
}
\description{
Assess a package for an up-to-date NEWS file
}
\examples{
assess_license(pkg_ref("riskmetric"))
}
\seealso{
\code{\link{score.pkg_metric_license}}

Other \code{assess_*} functions: 
\code{\link{assess_export_help}()},
\code{\link{assess_has_news}()},
\code{\link{assess_news_current}()},
\code{\link{assess}()}
}
\concept{\code{assess_*} functions}
