% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_news_current.R
\name{assess_news_current}
\alias{assess_news_current}
\title{Assess a package for an up-to-date NEWS file}
\usage{
assess_news_current(x, ...)
}
\arguments{
\item{x}{a \code{pkg_ref} package reference object}

\item{...}{additional arguments passed on to S3 methods, rarely used}
}
\value{
a \code{pkg_metric} containing a logical vector indicating whether each discovered NEWS file is up-to-date
}
\description{
Assess a package for an up-to-date NEWS file
}
\examples{
\dontrun{
assess_news_current(pkg_ref("riskmetric"))
}
}
\seealso{
\code{\link{metric_score.pkg_metric_news_current}}
}
