% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_bug_reports.R, R/assess_news_current.R
\name{assess_news_current}
\alias{assess_news_current}
\title{Assess a package for an up-to-date NEWS file}
\usage{
assess_news_current(x, ...)

assess_news_current(x, ...)
}
\arguments{
\item{x}{a \code{pkg_ref} package reference object}

\item{...}{additional arguments passed on to S3 methods, rarely used}

\item{x}{a \code{pkg_ref} package reference object}

\item{...}{additional arguments passed on to S3 methods, rarely used}
}
\value{
a \code{pkg_metric} containing a logical vector indicating whether each discovered NEWS file is up-to-date

a \code{pkg_metric} containing a logical vector indicating whether each discovered NEWS file is up-to-date
}
\description{
Assess a package for an up-to-date NEWS file

Assess a package for an up-to-date NEWS file
}
\examples{
assess_news_current(pkg_ref("riskmetric"))
assess_news_current(pkg_ref("riskmetric"))
}
\seealso{
\code{\link{score.pkg_metric_news_current}}

\code{\link{score.pkg_metric_news_current}}

Other \code{assess_*} functions: \code{\link{assess_export_help}},
  \code{\link{assess_has_news}}

Other \code{assess_*} functions: \code{\link{assess_export_help}},
  \code{\link{assess_has_news}}
}
\concept{\code{assess_*} functions}
