% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_r_cmd_check.R
\name{assess_r_cmd_check}
\alias{assess_r_cmd_check}
\title{Assess a package's results from running R CMD check}
\usage{
assess_r_cmd_check(x, ...)
}
\arguments{
\item{x}{a \code{pkg_ref} package reference object}

\item{...}{additional arguments passed on to S3 methods, rarely used}
}
\value{
a \code{pkg_metric} containing Tally of errors, warnings and notes from running R CMD check locally
}
\description{
Assess a package's results from running R CMD check
}
\examples{
\dontrun{
assess_r_cmd_check(pkg_ref("riskmetric"))
}
}
\seealso{
\code{\link{metric_score.pkg_metric_r_cmd_check}}

Other \code{assess_*} functions: 
\code{\link{assess_covr_coverage}()},
\code{\link{assess_downloads_1yr}()},
\code{\link{assess_export_help}()},
\code{\link{assess_has_bug_reports_url}()},
\code{\link{assess_has_maintainer}()},
\code{\link{assess_has_news}()},
\code{\link{assess_has_source_control}()},
\code{\link{assess_has_vignettes}()},
\code{\link{assess_has_website}()},
\code{\link{assess_last_30_bugs_status}()},
\code{\link{assess_license}()},
\code{\link{assess_news_current}()},
\code{\link{assess_remote_checks}()},
\code{\link{pkg_assess}()}
}
\concept{\code{assess_*} functions}
