% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_reverse_dependencies.R
\name{assess_reverse_dependencies}
\alias{assess_reverse_dependencies}
\title{Generate list of Reverse Dependencies for a package}
\usage{
assess_reverse_dependencies(x, ...)
}
\description{
Generate list of Reverse Dependencies for a package
}
\details{
The more packages that depend on a package the more chance
for errors/bugs to be found
}
