% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_security.R
\name{assess_security}
\alias{assess_security}
\title{Assess a package for known security vulnerabilities in the OSS Index}
\usage{
assess_security(x, ...)
}
\description{
Assess a package for known security vulnerabilities in the OSS Index
}
