% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_metric_error.R
\name{assessment_error_as_warning}
\alias{assessment_error_as_warning}
\title{Error handler for assessments to deescalate errors to warnings}
\usage{
assessment_error_as_warning(e, name, assessment)
}
\arguments{
\item{e}{an error raised during a package reference assessment}

\item{name}{the name of the package whose package reference assessment raised
the error}

\item{assessment}{the name of the assessment function which raised the error}
}
\value{
a pkg_metric object of pkg_metric_error subclass
}
\description{
Error handler for assessments to deescalate errors to warnings
}
\seealso{
Other assessment error handlers: \code{\link{assessment_error_empty}},
  \code{\link{assessment_error_throw}}
}
\concept{assessment error handlers}
