% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_metric_error.R
\name{assessment_error_empty}
\alias{assessment_error_empty}
\title{Error handler for assessments with safe fallback}
\usage{
assessment_error_empty(e)
}
\description{
Error handler for assessments with safe fallback
}
