% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_metric_error.R
\name{assessment_error_throw}
\alias{assessment_error_throw}
\title{Error handler for assessments to throw error immediately}
\usage{
assessment_error_throw(e, name, assessment)
}
\description{
Error handler for assessments to throw error immediately
}
