% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_class_extract.R
\name{bare_env}
\alias{bare_env}
\title{evaluate an expression with a  pkg_ref object reclassed as a bare environment
object, used to sidestep pkg_ref assignment guardrails}
\usage{
bare_env(x, expr, envir = parent.frame())
}
\description{
evaluate an expression with a  pkg_ref object reclassed as a bare environment
object, used to sidestep pkg_ref assignment guardrails
}
