% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_bug_reports.R
\name{bug_report_metadata}
\alias{bug_report_metadata}
\title{Helper for structuring bug reports}
\usage{
bug_report_metadata(bug_reports_data, x)
}
\arguments{
\item{bug_reports_data}{data to represent a bug report history - generally a
return object from making a request to a repository's issues API}

\item{x}{a \code{pkg_ref} object where a \code{bug_reports_host} field can be
found}
}
\description{
Helper for structuring bug reports
}
\keyword{internal}
