% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_ref_class.R
\name{cohort_ref}
\alias{cohort_ref}
\title{Cohort ref class}
\usage{
cohort_ref(x, library, ...)
}
\description{
This is a class to hold a list of pkg_refs along with extra
information about the environment the packages are to be used/installed in.
}
\concept{cohort_ref}
