% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_score.R
\name{cohort_score}
\alias{cohort_score}
\title{Score a cohort assessment, collapsing results into a single numeric
cohort_score() calculates the risk involved with using a cohort of packages. Risk ranges
from 0 (low-risk) to 1 (high-risk).}
\usage{
cohort_score(x, ..., error_handler = score_error_default)
}
\description{
Score a cohort assessment, collapsing results into a single numeric
cohort_score() calculates the risk involved with using a cohort of packages. Risk ranges
from 0 (low-risk) to 1 (high-risk).
}
