% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_examples.R
\name{examples_from_dir}
\alias{examples_from_dir}
\title{Build a List of Examples For Exported Objects Discovered Within a Given Directory}
\usage{
examples_from_dir(x)
}
\arguments{
\item{path}{a package directory path expected to contain exported objects}
}
\value{
a vector of parsed examples
}
\description{
Build a List of Examples For Exported Objects Discovered Within a Given Directory
}
\keyword{internal}
