% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_assess.R
\name{get_assessment_columns}
\alias{get_assessment_columns}
\title{Helper for retrieving a list of columns which contain pkg_metric objects}
\usage{
get_assessment_columns(tbl)
}
\arguments{
\item{tbl}{a \code{\link[tibble]{tibble}} to select columns among}
}
\value{
a logical vector of \code{pkg_metric} column indices
}
\description{
Helper for retrieving a list of columns which contain pkg_metric objects
}
