% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_memoised.R
\name{memoise_bioc_mirrors}
\alias{memoise_bioc_mirrors}
\title{Fetch BioC Mirrors Info}
\usage{
memoise_bioc_mirrors()
}
\value{
a data frame with mirror information
}
\description{
taken from utils::chooseBioCmirror
}
\keyword{internal}
