% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_covr_coverage.R
\name{metric_score.pkg_metric_covr_coverage}
\alias{metric_score.pkg_metric_covr_coverage}
\title{Score a package for unit test coverage}
\usage{
\method{metric_score}{pkg_metric_covr_coverage}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_covr_coverage} packge metric object}

\item{...}{additional arguments unused}
}
\value{
A \code{numeric}
}
\description{
Returns the overall test coverage from a covr coverage report
}
\examples{
metric_score(assess_covr_coverage(pkg_ref("riskmetric")))
}
\seealso{
Other \code{metric_score.*} functions: 
\code{\link{metric_score.pkg_metric_downloads_1yr}()},
\code{\link{metric_score.pkg_metric_export_help}()},
\code{\link{metric_score.pkg_metric_has_bug_reports_url}()},
\code{\link{metric_score.pkg_metric_has_maintainer}()},
\code{\link{metric_score.pkg_metric_has_news}()},
\code{\link{metric_score.pkg_metric_has_source_control}()},
\code{\link{metric_score.pkg_metric_has_vignettes}()},
\code{\link{metric_score.pkg_metric_has_website}()},
\code{\link{metric_score.pkg_metric_last_30_bugs_status}()},
\code{\link{metric_score.pkg_metric_license}()},
\code{\link{metric_score.pkg_metric_news_current}()},
\code{\link{metric_score.pkg_metric_r_cmd_check}()}
}
\concept{\code{metric_score.*} functions}
