% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_dependencies.R
\name{metric_score.pkg_metric_dependencies}
\alias{metric_score.pkg_metric_dependencies}
\title{Score a package for dependencies}
\usage{
\method{metric_score}{pkg_metric_dependencies}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_dependencies} packge metric object}

\item{...}{additional arguments unused}
}
\value{
A \code{numeric}
}
\description{
Returns the total number dependencies
}
\examples{
\dontrun{metric_score(assess_dependencies(pkg_ref("riskmetric")))
}
}
