% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_export_help.R
\name{metric_score.pkg_metric_export_help}
\alias{metric_score.pkg_metric_export_help}
\title{Score a package for availability of documentation for exported values}
\usage{
\method{metric_score}{pkg_metric_export_help}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_export_help} packge metric object}

\item{...}{additional arguments unused}
}
\value{
\code{1} if any NEWS files are found, otherwise \code{0}
}
\description{
Coerce a logical vector indicating availability of export documentation
}
\examples{
\dontrun{metric_score(assess_export_help(pkg_ref("riskmetric")))
}
}
