% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_bug_reports_url.R
\name{metric_score.pkg_metric_has_bug_reports_url}
\alias{metric_score.pkg_metric_has_bug_reports_url}
\title{Score a package for the presence of a bug report url}
\usage{
\method{metric_score}{pkg_metric_has_bug_reports_url}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_has_bug_reports_url} packge metric object}

\item{...}{additional arguments unused}
}
\value{
A logical value indicating whether the package has a BugReports field
  filled in
}
\description{
Score a package for the presence of a bug report url
}
\examples{
metric_score(assess_has_bug_reports_url(pkg_ref("riskmetric")))
}
\seealso{
Other \code{metric_score.*} functions:
\code{\link{metric_score.pkg_metric_covr_coverage}()},
\code{\link{metric_score.pkg_metric_downloads_1yr}()},
\code{\link{metric_score.pkg_metric_export_help}()},
\code{\link{metric_score.pkg_metric_has_maintainer}()},
\code{\link{metric_score.pkg_metric_has_news}()},
\code{\link{metric_score.pkg_metric_has_source_control}()},
\code{\link{metric_score.pkg_metric_has_vignettes}()},
\code{\link{metric_score.pkg_metric_has_website}()},
\code{\link{metric_score.pkg_metric_last_30_bugs_status}()},
\code{\link{metric_score.pkg_metric_license}()},
\code{\link{metric_score.pkg_metric_news_current}()},
\code{\link{metric_score.pkg_metric_r_cmd_check}()},
\code{\link{metric_score.pkg_metric_remote_checks}()}
}
\concept{\code{metric_score.*} functions}
