% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_has_maintainer.R
\name{metric_score.pkg_metric_has_maintainer}
\alias{metric_score.pkg_metric_has_maintainer}
\title{Score a package for inclusion of an associated maintainer}
\usage{
\method{metric_score}{pkg_metric_has_maintainer}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_has_maintainer} packge metric object}

\item{...}{additional arguments unused}
}
\value{
\code{1} if any maintainer is provided, otherwise \code{0}
}
\description{
Coerce a list of maintainers into a numeric value indicating whether the
number of listed maintainers is greater than 0.
}
\examples{
\dontrun{metric_score(assess_has_maintainer(pkg_ref("riskmetric")))
}
}
