% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_last_30_bugs_status.R
\name{metric_score.pkg_metric_last_30_bugs_status}
\alias{metric_score.pkg_metric_last_30_bugs_status}
\title{Score a package for number of recently opened BugReports that are now closed}
\usage{
\method{metric_score}{pkg_metric_last_30_bugs_status}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_last_30_bugs_status} packge metric object}

\item{...}{additional arguments unused}
}
\value{
a fractional value indicating percentage of last 30 bug reports that
  are now closed
}
\description{
Score a package for number of recently opened BugReports that are now closed
}
\examples{
\dontrun{metric_score(assess_last_30_bugs_status(pkg_ref("riskmetric")))
}
}
\seealso{
Other \code{metric_score.*} functions: 
\code{\link{metric_score.pkg_metric_covr_coverage}()},
\code{\link{metric_score.pkg_metric_downloads_1yr}()},
\code{\link{metric_score.pkg_metric_export_help}()},
\code{\link{metric_score.pkg_metric_has_bug_reports_url}()},
\code{\link{metric_score.pkg_metric_has_maintainer}()},
\code{\link{metric_score.pkg_metric_has_news}()},
\code{\link{metric_score.pkg_metric_has_source_control}()},
\code{\link{metric_score.pkg_metric_has_vignettes}()},
\code{\link{metric_score.pkg_metric_has_website}()},
\code{\link{metric_score.pkg_metric_license}()},
\code{\link{metric_score.pkg_metric_news_current}()}
}
\concept{\code{metric_score.*} functions}
