% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_license.R
\name{metric_score.pkg_metric_license}
\alias{metric_score.pkg_metric_license}
\title{Score a package for acceptable license}
\usage{
\method{metric_score}{pkg_metric_license}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_license} packge metric object}

\item{...}{additional arguments unused}
}
\value{
score of metric license
}
\description{
Maps a license string to a score
}
\examples{
metric_score(assess_license(pkg_ref("riskmetric")))
}
