% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_r_cmd_check.R
\name{metric_score.pkg_metric_r_cmd_check}
\alias{metric_score.pkg_metric_r_cmd_check}
\title{Score a package based on R CMD check results run locally}
\usage{
\method{metric_score}{pkg_metric_r_cmd_check}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_r_cmd_check} packge metric object}

\item{...}{additional arguments unused}
}
\value{
A weighted sum of errors and warnings of all tests preformed
}
\description{
The scoring function is
}
\examples{
\dontrun{metric_score(assess_r_cmd_check(pkg_ref("riskmetric")))
}
}
\seealso{
Other \code{metric_score.*} functions: 
\code{\link{metric_score.pkg_metric_covr_coverage}()},
\code{\link{metric_score.pkg_metric_downloads_1yr}()},
\code{\link{metric_score.pkg_metric_export_help}()},
\code{\link{metric_score.pkg_metric_has_bug_reports_url}()},
\code{\link{metric_score.pkg_metric_has_maintainer}()},
\code{\link{metric_score.pkg_metric_has_news}()},
\code{\link{metric_score.pkg_metric_has_source_control}()},
\code{\link{metric_score.pkg_metric_has_vignettes}()},
\code{\link{metric_score.pkg_metric_has_website}()},
\code{\link{metric_score.pkg_metric_last_30_bugs_status}()},
\code{\link{metric_score.pkg_metric_license}()},
\code{\link{metric_score.pkg_metric_news_current}()}
}
\concept{\code{metric_score.*} functions}
