% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_remote_checks.R
\name{metric_score.pkg_metric_remote_checks}
\alias{metric_score.pkg_metric_remote_checks}
\title{Score a package based on R CMD check results run by BioC or CRAN}
\usage{
\method{metric_score}{pkg_metric_remote_checks}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_remote_checks} packge metric object}

\item{...}{additional arguments unused}
}
\value{
a fractional value indicating percentage OS flavors that did not produce an error or warning from R CMD check
}
\description{
The scoring function is the number of OS flavors that passed with OK or NOTES + 0.5*the number of OS's that produced WARNINGS divided by the number of OS's checked
}
\examples{
\dontrun{metric_score(assess_remote_checks(pkg_ref("riskmetric")))
}
}
