% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_remote_checks.R
\name{metric_score.pkg_metric_remote_checks}
\alias{metric_score.pkg_metric_remote_checks}
\title{Score a package based on R CMD check results run by BioC or CRAN}
\usage{
\method{metric_score}{pkg_metric_remote_checks}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_remote_checks} packge metric object}

\item{...}{additional arguments unused}
}
\value{
a fractional value indicating percentage OS flavors that did not produce an error or warning from R CMD check
}
\description{
Score a package based on R CMD check results run by BioC or CRAN
}
\examples{
\dontrun{metric_score(assess_remote_checks(pkg_ref("riskmetric")))
}
}
\seealso{
Other \code{metric_score.*} functions: \code{\link{metric_score.pkg_metric_covr_coverage}},
  \code{\link{metric_score.pkg_metric_downloads_1yr}},
  \code{\link{metric_score.pkg_metric_export_help}},
  \code{\link{metric_score.pkg_metric_has_bug_reports_url}},
  \code{\link{metric_score.pkg_metric_has_maintainer}},
  \code{\link{metric_score.pkg_metric_has_news}},
  \code{\link{metric_score.pkg_metric_has_source_control}},
  \code{\link{metric_score.pkg_metric_has_vignettes}},
  \code{\link{metric_score.pkg_metric_has_website}},
  \code{\link{metric_score.pkg_metric_last_30_bugs_status}},
  \code{\link{metric_score.pkg_metric_license}},
  \code{\link{metric_score.pkg_metric_news_current}}
}
\concept{\code{metric_score.*} functions}
