% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_reverse_dependencies.R
\name{metric_score.pkg_metric_reverse_dependencies}
\alias{metric_score.pkg_metric_reverse_dependencies}
\title{Scoring method for number of reverse dependencies a package has}
\usage{
\method{metric_score}{pkg_metric_reverse_dependencies}(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric_reverse_dependencies} packge metric object}

\item{...}{additional arguments unused}
}
\value{
The count of reverse dependencies a package has
}
\description{
Scoring method for number of reverse dependencies a package has
}
\examples{
\dontrun{metric_score(assess_reverse_dependencies(pkg_ref("riskmetric")))
}
}
