% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_metric.R
\name{pkg_metric}
\alias{pkg_metric}
\title{A helper for structuring assessment return objects for dispatch with the
score function}
\usage{
pkg_metric(x, ..., class = c())
}
\arguments{
\item{x}{data to store as a \code{pkg_metric}}

\item{...}{additional attributes to bind to the \code{pkg_metric} object}

\item{class}{a subclass to differentiate the \code{pkg_metric} object}
}
\value{
a \code{pkg_metric} object
}
\description{
A helper for structuring assessment return objects for dispatch with the
score function
}
