% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_help_aliases.R
\name{pkg_ref_cache.help_aliases}
\alias{pkg_ref_cache.help_aliases}
\title{Cache a character vector mapping exported values to documentation filenames}
\usage{
pkg_ref_cache.help_aliases(x, name, ...)
}
\arguments{
\item{x}{a package reference object}

\item{name}{the name of the field that needs to be cached}

\item{...}{additional arguments used for computing cached values}
}
\description{
Cache a character vector mapping exported values to documentation filenames
}
\seealso{
Other package reference cache: 
\code{\link{pkg_ref_cache.bug_reports_host}()},
\code{\link{pkg_ref_cache.bug_reports_url}()},
\code{\link{pkg_ref_cache.bug_reports}()},
\code{\link{pkg_ref_cache.description}()},
\code{\link{pkg_ref_cache.downloads}()},
\code{\link{pkg_ref_cache.help}()},
\code{\link{pkg_ref_cache.news_urls}()},
\code{\link{pkg_ref_cache.news}()},
\code{\link{pkg_ref_cache.repo_base_url}()},
\code{\link{pkg_ref_cache.tarball_url}()},
\code{\link{pkg_ref_cache.vignettes}()},
\code{\link{pkg_ref_cache.web_html}()},
\code{\link{pkg_ref_cache.web_url}()},
\code{\link{pkg_ref_cache}()}
}
\concept{package reference cache}
