% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_web_html.R
\name{pkg_ref_cache.web_html}
\alias{pkg_ref_cache.web_html}
\title{Cache package's remote display page HTML}
\usage{
pkg_ref_cache.web_html(x, name, ...)
}
\arguments{
\item{x}{a package reference object}

\item{name}{the name of the field that needs to be cached}

\item{...}{additional arguments used for computing cached values}
}
\description{
Cache package's remote display page HTML
}
\seealso{
Other package reference cache: 
\code{\link{pkg_ref_cache.description}()},
\code{\link{pkg_ref_cache.help_aliases}()},
\code{\link{pkg_ref_cache.help}()},
\code{\link{pkg_ref_cache.news_urls}()},
\code{\link{pkg_ref_cache.news}()},
\code{\link{pkg_ref_cache.repo_base_url}()},
\code{\link{pkg_ref_cache.tarball_url}()},
\code{\link{pkg_ref_cache.web_url}()},
\code{\link{pkg_ref_cache}()}
}
\concept{package reference cache}
