% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_behaviors.R
\name{require_cache_behaviors}
\alias{require_cache_behaviors}
\title{Stop if a function requires disabled behaviors}
\usage{
require_cache_behaviors(behaviors)
}
\arguments{
\item{behaviors}{a character vector of behavior flags to assert as
requirements for metadata caching. values must have an entry found in
riskmetric:::cache_behaviors list}
}
\value{
a boolean value
}
\description{
Stop if a function requires disabled behaviors
}
\keyword{internal}
