% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskmetric.R
\docType{package}
\name{riskmetric}
\alias{riskmetric}
\alias{riskmetric-package}
\title{riskmetric}
\description{
Facilities for assessing R packages against a number of metrics to help
quantify their robustness.
}
