% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_assess.R
\name{roxygen_assess_family}
\alias{roxygen_assess_family}
\title{Helper for creating a roxygen header from template for assess_* functions}
\usage{
roxygen_assess_family(
  name,
  return_type = "an atomic assessment result",
  dontrun = FALSE
)
}
\arguments{
\item{name}{the name of the assessment, assuming naming conventions are
followed}

\item{return_type}{an optional added commentary about the return type of the
assessment function}

\item{dontrun}{logical indicating whether examples should be wrapped in
a dontrun block. This is particularly useful for assessments which may
require an internet connection.}
}
\value{
roxygen section template for assess family functions
}
\description{
Helper for creating a roxygen header from template for assess_* functions
}
\examples{
\dontrun{
#' @eval roxygen_assess_family(
#'   "has_news",
#'   "an integer value indicating the number of discovered NEWS files")
}

}
