% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache__behaviors.R
\name{roxygen_cache_behaviors}
\alias{roxygen_cache_behaviors}
\title{Helper for documenting both declare_cache_behavior parameters and options
list}
\usage{
roxygen_cache_behaviors(fmt = "\%s: \%s", name_fmt = "\%s",
  annotation_fmt = "\%s", wrap_fmt = "\%s", collapse = "\\n")
}
\description{
Helper for documenting both declare_cache_behavior parameters and options
list
}
