% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\alias{score.default}
\alias{score.tbl_df}
\title{Score a package assement, collapsing results into a single numeric}
\usage{
score(x, ..., envir = parent.frame())

\method{score}{default}(x, ..., envir = parent.frame())

\method{score}{tbl_df}(x, ..., envir = parent.frame())
}
\description{
Score a package assement, collapsing results into a single numeric
}
