% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Score a package assement, collapsing results into a single numeric}
\usage{
score(x, ..., error_handler = score_error_default)
}
\arguments{
\item{x}{a \code{pkg_metric} object, whose subclass is used to choose the
appropriate scoring method for the atomic metric metadata. Optionally, a
\code{\link[tibble]{tibble}} can be provided, in which cases all
\code{pkg_metric} values will be scored.}

\item{...}{additional arguments unused}

\item{error_handler}{specify a function to be called if the class can't be
identified. Most commonly this occurs for \code{pkg_metric} objects of
subclass \code{pkg_metric_error}, which is produced when an error is
encountered when calculating an associated assessment.}
}
\value{
a numeric value if a single \code{pkg_metric} is provided, or a
  \code{\link[tibble]{tibble}} with \code{pkg_metric} objects scored and
  returned as numeric values when a \code{\link[tibble]{tibble}} is provided.
}
\description{
Scores reflect the adherence to best practices. Scores always range from 0
(worst-) to 1 (best- practice), and can then be used to consistently evaluate
the risk involved with using a package.
}
\examples{
# scoring a single assessment
score(assess_has_news(pkg_ref("riskmetric")))

# scoring many assessments as a tibble
\dontrun{
library(dplyr)
score(assess(as_tibble(pkg_ref(c("riskmetric", "riskmetric")))))
}

}
\seealso{
score_error_default score_error_zero score_error_NA

<<<<<<< HEAD
Other \code{score.*} functions: \code{\link{score.pkg_metric_export_help}},
  \code{\link{score.pkg_metric_has_news}},
  \code{\link{score.pkg_metric_last_30_bugs_status}},
  \code{\link{score.pkg_metric_news_current}}
=======
Other \code{score.*} functions: 
\code{\link{score.pkg_metric_downloads_1yr}()},
\code{\link{score.pkg_metric_export_help}()},
\code{\link{score.pkg_metric_has_news}()},
\code{\link{score.pkg_metric_license}()},
\code{\link{score.pkg_metric_news_current}()}
>>>>>>> master
}
\concept{\code{score.*} functions}
