% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_error_default}
\alias{score_error_default}
\title{Default score error handling, emitting a warning and returning 0}
\usage{
score_error_default(x, ...)
}
\arguments{
\item{x}{a \code{pkg_metric} object, whose subclass is used to choose the
appropriate scoring method for the atomic metric metadata. Optionally, a
\code{\link[tibble]{tibble}} can be provided, in which cases all
\code{pkg_metric} values will be scored.}

\item{...}{additional arguments unused}
}
\description{
Default score error handling, emitting a warning and returning 0
}
\seealso{
Other \code{score_error_*} functions: \code{\link{score_error_NA}},
  \code{\link{score_error_zero}}
}
\concept{\code{score_error_*} functions}
