% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_score.R
\name{score_error_NA}
\alias{score_error_NA}
\title{Score error handler to silently return NA}
\usage{
score_error_NA(...)
}
\arguments{
\item{...}{Additional arguments unused}
}
\value{
a value of package score
}
\description{
Score error handler to silently return NA
}
