% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_score.R
\name{score_error_zero}
\alias{score_error_zero}
\title{Score error handler to silently return 0}
\usage{
score_error_zero(...)
}
\arguments{
\item{...}{Additional arguments unused}
}
\value{
a value of package score
}
\description{
Score error handler to silently return 0
}
