% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_assess.R
\name{use_assessments_column_names}
\alias{use_assessments_column_names}
\title{reassign assignment list names with column_name attribute if available}
\usage{
use_assessments_column_names(x)
}
\arguments{
\item{x}{list of columns for which to consider friendly column name
attributes}
}
\value{
a vector of friendly column names if available
}
\description{
reassign assignment list names with column_name attribute if available
}
\keyword{internal}
