% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ref_cache_vignettes.R
\name{vignettes_from_url}
\alias{vignettes_from_url}
\title{Build a List of Vignettes Files Discovered Within a Package Website}
\usage{
vignettes_from_url(web_html)
}
\arguments{
\item{web_html}{a \code{pkg_ref$web_html} document containing links to Vignettes files}
}
\value{
a vector of Vignettes files
}
\description{
Build a List of Vignettes Files Discovered Within a Package Website
}
